refactor: split ingest_test.go into per-mode test files
- ingest_helpers_test.go: shared vars, consts, helper functions
- ingest_backfill_test.go: 10 historical backfill tests
- ingest_catchup_test.go: 7 catchup mode tests
- ingest_test.go: 8 shared/core tests (down from 2972 to 831 lines)

Also removed unused txHash1/txHash2 constants (deadcode).
