+++ internal/services/ingest_test.go
@@ -4,15 +4,10 @@ import (
 	"context"
 	"fmt"
 	"testing"
-	"time"
 
-	"github.com/google/uuid"
 	"github.com/jackc/pgx/v5"
-	"github.com/jackc/pgx/v5/pgxpool"
 	"github.com/stellar/go-stellar-sdk/ingest/ledgerbackend"
-	"github.com/stellar/go-stellar-sdk/keypair"
 	"github.com/stellar/go-stellar-sdk/network"
-	"github.com/stellar/go-stellar-sdk/toid"
 	"github.com/stellar/go-stellar-sdk/xdr"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/mock"
@@ -26,36 +21,6 @@ import (
 	"github.com/stellar/wallet-backend/internal/indexer/types"
 	"github.com/stellar/wallet-backend/internal/metrics"
 	"github.com/stellar/wallet-backend/internal/signing/store"
-)
-
-var (
-	testKP1   = keypair.MustRandom()
-	testKP2   = keypair.MustRandom()
-	testAddr1 = testKP1.Address()
-	testAddr2 = testKP2.Address()
-)
-
-const (
-	defaultGetLedgersLimit = 50
-
-	// Test hash constants for ingest tests (64-char hex strings for BYTEA storage)
-	flushTxHash1   = "f1f1f1f1f1f1f1f1f1f1f1f1f1f1f1f1f1f1f1f1f1f1f1f1f1f1f1f1f1f1f101"
-	flushTxHash2   = "f2f2f2f2f2f2f2f2f2f2f2f2f2f2f2f2f2f2f2f2f2f2f2f2f2f2f2f2f2f2f202"
-	flushTxHash3   = "f3f3f3f3f3f3f3f3f3f3f3f3f3f3f3f3f3f3f3f3f3f3f3f3f3f3f3f3f3f3f303"
-	flushTxHash4   = "f4f4f4f4f4f4f4f4f4f4f4f4f4f4f4f4f4f4f4f4f4f4f4f4f4f4f4f4f4f4f404"
-	catchupTxHash1 = "c1c1c1c1c1c1c1c1c1c1c1c1c1c1c1c1c1c1c1c1c1c1c1c1c1c1c1c1c1c1c101"
-	catchupTxHash2 = "c2c2c2c2c2c2c2c2c2c2c2c2c2c2c2c2c2c2c2c2c2c2c2c2c2c2c2c2c2c2c202"
-	catchupTxHash3 = "c3c3c3c3c3c3c3c3c3c3c3c3c3c3c3c3c3c3c3c3c3c3c3c3c3c3c3c3c3c3c303"
-	catchupTxHash4 = "c4c4c4c4c4c4c4c4c4c4c4c4c4c4c4c4c4c4c4c4c4c4c4c4c4c4c4c4c4c4c404"
-	catchupTxHash5 = "c5c5c5c5c5c5c5c5c5c5c5c5c5c5c5c5c5c5c5c5c5c5c5c5c5c5c5c5c5c5c505"
-	catchupTxHash6 = "c6c6c6c6c6c6c6c6c6c6c6c6c6c6c6c6c6c6c6c6c6c6c6c6c6c6c6c6c6c6c606"
-	prevTxHash     = "aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa"
-	txHash1        = "1111111111111111111111111111111111111111111111111111111111111111"
-	txHash2        = "2222222222222222222222222222222222222222222222222222222222222222"
-
-	// Test fixtures for ledger metadata
-	ledgerMetadataWith0Tx = "AAAAAQAAAACB7Zh2o0NTFwl1nvs7xr3SJ7w8PpwnSRb8QyG9k6acEwAAABaeASPlzu/ZFxwyyWsxtGoj3KCrybm2yN7WOweR0BWdLYjyoO5BI41g1PFT+iHW68giP49Koo+q3VmH8I4GdtW2AAAAAGhTTB8AAAAAAAAAAQAAAAC1XRCyu30oTtXAOkel4bWQyQ9Xg1VHHMRQe76CBNI8iwAAAEDSH4sE7cL7UJyOqUo9ZZeNqPT7pt7su8iijHjWYg4MbeFUh/gkGf6N40bZjP/dlIuGXmuEhWoEX0VTV58xOB4C3z9hmASpL9tAVxktxD3XSOp3itxSvEmM6AUkwBS4ERm+pITz+1V1m+3/v6eaEKglCnon3a5xkn02sLltJ9CSzwAAEYIN4Lazp2QAAAAAAAMtYtQzAAAAAAAAAAAAAAAMAAAAZABMS0AAAADIXukLfWC53MCmzxKd/+LBbaYxQkgxATFDLI3hWj7EqWgAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAGeASPlzu/ZFxwyyWsxtGoj3KCrybm2yN7WOweR0BWdLQAAAAIAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAA9yHMAAAAAAAAAAA=="
-	ledgerMetadataWith1Tx = "AAAAAQAAAAD8G2qemHnBKFkbq90RTagxAypNnA7DXDc63Giipq9mNwAAABYLEZ5DrTv6njXTOAFEdOO0yeLtJjCRyH4ryJkgpRh7VPJvwbisrc9A0yzFxxCdkICgB3Gv7qHOi8ZdsK2CNks2AAAAAGhTTAsAAAAAAAAAAQAAAACoJM0YvJ11Bk0pmltbrKQ7w6ovMmk4FT2ML5u1y23wMwAAAEAunZtorOSbnRpgnykoDe4kzAvLwNXefncy1R/1ynBWyDv0DfdnqJ6Hcy/0AJf6DkBZlRayg775h3HjV0GKF/oPua7l8wkLlJBtSk1kRDt55qSf6btSrgcupB/8bnpJfUUgZJ76saUrj29HukYHS1bq7SyuoCAY+5F9iBYTmW1G9QAAEX4N4Lazp2QAAAAAAAMtS3veAAAAAAAAAAAAAAAMAAAAZABMS0AAAADIXukLfWC53MCmzxKd/+LBbaYxQkgxATFDLI3hWj7EqWgAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAELEZ5DrTv6njXTOAFEdOO0yeLtJjCRyH4ryJkgpRh7VAAAAAIAAAAAAAAAAQAAAAAAAAABAAAAAAAAAGQAAAABAAAAAgAAAADg4mtiLKjJVgrmOpO9+Ff3XAmnycHyNUKu/v9KhHevAAAAAGQAAA7FAAAAGgAAAAAAAAAAAAAAAQAAAAAAAAABAAAAALvqzdVyRxgBMcLzbw1wNWcJYHPNPok1GdVSgmy4sjR2AAAAAVVTREMAAAAA4OJrYiyoyVYK5jqTvfhX91wJp8nB8jVCrv7/SoR3rwAAAAACVAvkAAAAAAAAAAABhHevAAAAAEDq2yIDzXUoLboBHQkbr8U2oKqLzf0gfpwXbmRPLB6Ek3G8uCEYyry1vt5Sb+LCEd81fefFQcQN0nydr1FmiXcDAAAAAAAAAAAAAAABXFSiWcxpDRa8frBs1wbEaMUw4hMe7ctFtdw3Ci73IEwAAAAAAAAAZAAAAAAAAAABAAAAAAAAAAEAAAAAAAAAAAAAAAIAAAADAAARfQAAAAAAAAAA4OJrYiyoyVYK5jqTvfhX91wJp8nB8jVCrv7/SoR3rwAAAAAukO3GPAAADsUAAAAZAAAAAAAAAAAAAAAAAAAAAAEAAAAAAAAAAAAAAQAAAAAAAAAAAAAAAAAAAAAAAAACAAAAAAAAAAAAAAAAAAAAAwAAAAAAABF9AAAAAGhTTAYAAAAAAAAAAQAAEX4AAAAAAAAAAODia2IsqMlWCuY6k734V/dcCafJwfI1Qq7+/0qEd68AAAAALpDtxdgAAA7FAAAAGQAAAAAAAAAAAAAAAAAAAAABAAAAAAAAAAAAAAEAAAAAAAAAAAAAAAAAAAAAAAAAAgAAAAAAAAAAAAAAAAAAAAMAAAAAAAARfQAAAABoU0wGAAAAAAAAAAMAAAAAAAAAAgAAAAMAABF+AAAAAAAAAADg4mtiLKjJVgrmOpO9+Ff3XAmnycHyNUKu/v9KhHevAAAAAC6Q7cXYAAAOxQAAABkAAAAAAAAAAAAAAAAAAAAAAQAAAAAAAAAAAAABAAAAAAAAAAAAAAAAAAAAAAAAAAIAAAAAAAAAAAAAAAAAAAADAAAAAAAAEX0AAAAAaFNMBgAAAAAAAAABAAARfgAAAAAAAAAA4OJrYiyoyVYK5jqTvfhX91wJp8nB8jVCrv7/SoR3rwAAAAAukO3F2AAADsUAAAAaAAAAAAAAAAAAAAAAAAAAAAEAAAAAAAAAAAAAAQAAAAAAAAAAAAAAAAAAAAAAAAACAAAAAAAAAAAAAAAAAAAAAwAAAAAAABF+AAAAAGhTTAsAAAAAAAAAAQAAAAIAAAADAAARcwAAAAEAAAAAu+rN1XJHGAExwvNvDXA1Zwlgc80+iTUZ1VKCbLiyNHYAAAABVVNEQwAAAADg4mtiLKjJVgrmOpO9+Ff3XAmnycHyNUKu/v9KhHevAAAAAAlQL5AAf/////////8AAAABAAAAAAAAAAAAAAABAAARfgAAAAEAAAAAu+rN1XJHGAExwvNvDXA1Zwlgc80+iTUZ1VKCbLiyNHYAAAABVVNEQwAAAADg4mtiLKjJVgrmOpO9+Ff3XAmnycHyNUKu/v9KhHevAAAAAAukO3QAf/////////8AAAABAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAA8RxEAAAAAAAAAAA=="
 )
 
 func Test_generateAdvisoryLockID(t *testing.T) {
@@ -190,441 +155,6 @@ func Test_ingestService_splitGapsIntoBatches(t *testing.T) {
 		})
 	}
 }
-
-func Test_ingestService_calculateBackfillGaps(t *testing.T) {
-	dbt := dbtest.Open(t)
-	defer dbt.Close()
-	ctx := context.Background()
-	dbConnectionPool, err := db.OpenDBConnectionPool(ctx, dbt.DSN)
-	require.NoError(t, err)
-	defer dbConnectionPool.Close()
-
-	testCases := []struct {
-		name         string
-		startLedger  uint32
-		endLedger    uint32
-		setupDB      func(t *testing.T)
-		expectedGaps []data.LedgerRange
-	}{
-		{
-			name:        "entirely_before_oldest_ingested_ledger",
-			startLedger: 50,
-			endLedger:   80,
-			setupDB: func(t *testing.T) {
-				// Actual oldest ledger in transactions is 100
-				_, err := dbConnectionPool.Exec(ctx,
-					`INSERT INTO transactions (hash, to_id, envelope_xdr, fee_charged, result_code, meta_xdr, ledger_number, ledger_created_at)
-					VALUES ('anchor_hash', 1, 'env', 100, 'TransactionResultCodeTxSuccess', 'meta', 100, NOW())`)
-				require.NoError(t, err)
-			},
-			expectedGaps: []data.LedgerRange{
-				{GapStart: 50, GapEnd: 80},
-			},
-		},
-		{
-			name:        "overlaps_with_oldest_ingested_ledger",
-			startLedger: 50,
-			endLedger:   150,
-			setupDB: func(t *testing.T) {
-				// Insert transactions for 100-200 (no gaps); oldest is 100
-				for ledger := uint32(100); ledger <= 200; ledger++ {
-					_, err := dbConnectionPool.Exec(ctx,
-						`INSERT INTO transactions (hash, to_id, envelope_xdr, fee_charged, result_code, meta_xdr, ledger_number, ledger_created_at)
-						VALUES ($1, $2, 'env', 100, 'TransactionResultCodeTxSuccess', 'meta', $3, NOW())`,
-						"hash"+string(rune(ledger)), ledger, ledger)
-					require.NoError(t, err)
-				}
-			},
-			expectedGaps: []data.LedgerRange{
-				{GapStart: 50, GapEnd: 99},
-			},
-		},
-		{
-			name:        "entirely_within_ingested_range_no_gaps",
-			startLedger: 110,
-			endLedger:   150,
-			setupDB: func(t *testing.T) {
-				// Insert transactions for 100-200 (no gaps); oldest is 100
-				for ledger := uint32(100); ledger <= 200; ledger++ {
-					_, err := dbConnectionPool.Exec(ctx,
-						`INSERT INTO transactions (hash, to_id, envelope_xdr, fee_charged, result_code, meta_xdr, ledger_number, ledger_created_at)
-						VALUES ($1, $2, 'env', 100, 'TransactionResultCodeTxSuccess', 'meta', $3, NOW())`,
-						"hash"+string(rune(ledger)), ledger, ledger)
-					require.NoError(t, err)
-				}
-			},
-			expectedGaps: []data.LedgerRange{},
-		},
-		{
-			name:        "entirely_within_ingested_range_with_gaps",
-			startLedger: 110,
-			endLedger:   180,
-			setupDB: func(t *testing.T) {
-				// Insert transactions with gaps: 100-120, 150-200 (gap at 121-149); oldest is 100
-				for ledger := uint32(100); ledger <= 120; ledger++ {
-					_, err := dbConnectionPool.Exec(ctx,
-						`INSERT INTO transactions (hash, to_id, envelope_xdr, fee_charged, result_code, meta_xdr, ledger_number, ledger_created_at)
-						VALUES ($1, $2, 'env', 100, 'TransactionResultCodeTxSuccess', 'meta', $3, NOW())`,
-						"hash"+string(rune(ledger)), ledger, ledger)
-					require.NoError(t, err)
-				}
-				for ledger := uint32(150); ledger <= 200; ledger++ {
-					_, err := dbConnectionPool.Exec(ctx,
-						`INSERT INTO transactions (hash, to_id, envelope_xdr, fee_charged, result_code, meta_xdr, ledger_number, ledger_created_at)
-						VALUES ($1, $2, 'env', 100, 'TransactionResultCodeTxSuccess', 'meta', $3, NOW())`,
-						"hash"+string(rune(ledger)), ledger, ledger)
-					require.NoError(t, err)
-				}
-			},
-			expectedGaps: []data.LedgerRange{
-				{GapStart: 121, GapEnd: 149},
-			},
-		},
-		{
-			// After TimescaleDB retention drops old chunks, the stored oldest_ledger_cursor
-			// may point to a ledger that no longer exists in the transactions table.
-			// calculateBackfillGaps should ignore the cursor and use the actual oldest
-			// ledger from the transactions table.
-			name:        "stale_cursor_is_ignored_uses_actual_oldest_from_transactions",
-			startLedger: 50,
-			endLedger:   150,
-			setupDB: func(t *testing.T) {
-				// Cursor claims oldest is 50 (stale — retention dropped ledgers 50-99)
-				_, err := dbConnectionPool.Exec(ctx, `INSERT INTO ingest_store (key, value) VALUES ('oldest_ledger_cursor', 50)`)
-				require.NoError(t, err)
-				// Actual transactions only exist from 100 onwards
-				for ledger := uint32(100); ledger <= 200; ledger++ {
-					_, err := dbConnectionPool.Exec(ctx,
-						`INSERT INTO transactions (hash, to_id, envelope_xdr, fee_charged, result_code, meta_xdr, ledger_number, ledger_created_at)
-						VALUES ($1, $2, 'env', 100, 'TransactionResultCodeTxSuccess', 'meta', $3, NOW())`,
-						"hash"+string(rune(ledger)), ledger, ledger)
-					require.NoError(t, err)
-				}
-			},
-			// Oldest in transactions is 100, so gap [50,99] should be backfilled
-			expectedGaps: []data.LedgerRange{
-				{GapStart: 50, GapEnd: 99},
-			},
-		},
-	}
-
-	for _, tc := range testCases {
-		t.Run(tc.name, func(t *testing.T) {
-			// Clean up
-			_, err := dbConnectionPool.Exec(ctx, "DELETE FROM transactions")
-			require.NoError(t, err)
-			_, err = dbConnectionPool.Exec(ctx, "DELETE FROM ingest_store")
-			require.NoError(t, err)
-
-			mockMetricsService := metrics.NewMockMetricsService()
-			mockMetricsService.On("RegisterPoolMetrics", "ledger_indexer", mock.Anything).Return()
-			mockMetricsService.On("RegisterPoolMetrics", "backfill", mock.Anything).Return()
-			mockMetricsService.On("ObserveDBQueryDuration", mock.Anything, mock.Anything, mock.Anything).Return().Maybe()
-			mockMetricsService.On("IncDBQuery", mock.Anything, mock.Anything).Return().Maybe()
-			defer mockMetricsService.AssertExpectations(t)
-
-			models, err := data.NewModels(dbConnectionPool, mockMetricsService)
-			require.NoError(t, err)
-
-			if tc.setupDB != nil {
-				tc.setupDB(t)
-			}
-
-			mockAppTracker := apptracker.MockAppTracker{}
-			mockRPCService := RPCServiceMock{}
-			mockRPCService.On("NetworkPassphrase").Return(network.TestNetworkPassphrase).Maybe()
-			mockChAccStore := &store.ChannelAccountStoreMock{}
-			mockLedgerBackend := &LedgerBackendMock{}
-			mockArchive := &HistoryArchiveMock{}
-
-			svc, err := NewIngestService(IngestServiceConfig{
-				IngestionMode:          IngestionModeBackfill,
-				Models:                 models,
-				LatestLedgerCursorName: "latest_ledger_cursor",
-				OldestLedgerCursorName: "oldest_ledger_cursor",
-				AppTracker:             &mockAppTracker,
-				RPCService:             &mockRPCService,
-				LedgerBackend:          mockLedgerBackend,
-				ChannelAccountStore:    mockChAccStore,
-				MetricsService:         mockMetricsService,
-				GetLedgersLimit:        defaultGetLedgersLimit,
-				Network:                network.TestNetworkPassphrase,
-				NetworkPassphrase:      network.TestNetworkPassphrase,
-				Archive:                mockArchive,
-				SkipTxMeta:             false,
-				SkipTxEnvelope:         false,
-			})
-			require.NoError(t, err)
-
-			gaps, err := svc.calculateBackfillGaps(ctx, tc.startLedger, tc.endLedger)
-			require.NoError(t, err)
-			assert.Equal(t, tc.expectedGaps, gaps)
-		})
-	}
-}
-
-func Test_HistoricalBackfill_Validation(t *testing.T) {
-	dbt := dbtest.Open(t)
-	defer dbt.Close()
-	ctx := context.Background()
-	dbConnectionPool, err := db.OpenDBConnectionPool(ctx, dbt.DSN)
-	require.NoError(t, err)
-	defer dbConnectionPool.Close()
-
-	testCases := []struct {
-		name                  string
-		startLedger           uint32
-		endLedger             uint32
-		latestIngested        uint32
-		expectValidationError bool
-		errorContains         string
-	}{
-		{
-			name:                  "valid_range",
-			startLedger:           50,
-			endLedger:             80,
-			latestIngested:        100,
-			expectValidationError: false,
-		},
-		{
-			name:                  "end_exceeds_latest",
-			startLedger:           50,
-			endLedger:             150,
-			latestIngested:        100,
-			expectValidationError: true,
-			errorContains:         "end ledger 150 cannot be greater than latest ingested ledger 100",
-		},
-	}
-
-	for _, tc := range testCases {
-		t.Run(tc.name, func(t *testing.T) {
-			_, err := dbConnectionPool.Exec(ctx, "DELETE FROM transactions")
-			require.NoError(t, err)
-			_, err = dbConnectionPool.Exec(ctx, "DELETE FROM ingest_store")
-			require.NoError(t, err)
-
-			_, err = dbConnectionPool.Exec(ctx,
-				`INSERT INTO ingest_store (key, value) VALUES ('latest_ledger_cursor', $1)`,
-				fmt.Sprintf("%d", tc.latestIngested))
-			require.NoError(t, err)
-			_, err = dbConnectionPool.Exec(ctx,
-				`INSERT INTO ingest_store (key, value) VALUES ('oldest_ledger_cursor', $1)`,
-				fmt.Sprintf("%d", tc.latestIngested))
-			require.NoError(t, err)
-
-			mockMetricsService := metrics.NewMockMetricsService()
-			mockMetricsService.On("RegisterPoolMetrics", "ledger_indexer", mock.Anything).Return()
-			mockMetricsService.On("RegisterPoolMetrics", "backfill", mock.Anything).Return()
-			mockMetricsService.On("ObserveDBQueryDuration", mock.Anything, mock.Anything, mock.Anything).Return().Maybe()
-			mockMetricsService.On("IncDBQuery", mock.Anything, mock.Anything).Return().Maybe()
-			defer mockMetricsService.AssertExpectations(t)
-
-			models, err := data.NewModels(dbConnectionPool, mockMetricsService)
-			require.NoError(t, err)
-
-			mockBackendFactory := func(ctx context.Context) (ledgerbackend.LedgerBackend, error) {
-				return nil, fmt.Errorf("mock backend factory error")
-			}
-
-			svc, err := NewIngestService(IngestServiceConfig{
-				IngestionMode:          IngestionModeBackfill,
-				Models:                 models,
-				LatestLedgerCursorName: "latest_ledger_cursor",
-				OldestLedgerCursorName: "oldest_ledger_cursor",
-				AppTracker:             &apptracker.MockAppTracker{},
-				RPCService:             &RPCServiceMock{},
-				LedgerBackend:          &LedgerBackendMock{},
-				LedgerBackendFactory:   mockBackendFactory,
-				ChannelAccountStore:    &store.ChannelAccountStoreMock{},
-				MetricsService:         mockMetricsService,
-				GetLedgersLimit:        defaultGetLedgersLimit,
-				Network:                network.TestNetworkPassphrase,
-				NetworkPassphrase:      network.TestNetworkPassphrase,
-				Archive:                &HistoryArchiveMock{},
-				SkipTxMeta:             false,
-				SkipTxEnvelope:         false,
-				BackfillBatchSize:      100,
-			})
-			require.NoError(t, err)
-
-			err = svc.startHistoricalBackfill(ctx, tc.startLedger, tc.endLedger)
-			if tc.expectValidationError {
-				require.Error(t, err)
-				assert.Contains(t, err.Error(), tc.errorContains)
-			} else {
-				// Historical mode does not return error on batch failures, just logs them
-				require.NoError(t, err)
-			}
-		})
-	}
-}
-
-func Test_Catchup_Validation(t *testing.T) {
-	dbt := dbtest.Open(t)
-	defer dbt.Close()
-	ctx := context.Background()
-	dbConnectionPool, err := db.OpenDBConnectionPool(ctx, dbt.DSN)
-	require.NoError(t, err)
-	defer dbConnectionPool.Close()
-
-	testCases := []struct {
-		name                  string
-		startLedger           uint32
-		endLedger             uint32
-		latestIngested        uint32
-		expectValidationError bool
-		errorContains         string
-	}{
-		{
-			name:                  "valid_range",
-			startLedger:           101,
-			endLedger:             150,
-			latestIngested:        100,
-			expectValidationError: false,
-		},
-		{
-			name:                  "start_not_next_ledger",
-			startLedger:           105,
-			endLedger:             150,
-			latestIngested:        100,
-			expectValidationError: true,
-			errorContains:         "catchup must start from ledger 101",
-		},
-	}
-
-	for _, tc := range testCases {
-		t.Run(tc.name, func(t *testing.T) {
-			_, err := dbConnectionPool.Exec(ctx, "DELETE FROM transactions")
-			require.NoError(t, err)
-			_, err = dbConnectionPool.Exec(ctx, "DELETE FROM ingest_store")
-			require.NoError(t, err)
-
-			_, err = dbConnectionPool.Exec(ctx,
-				`INSERT INTO ingest_store (key, value) VALUES ('latest_ledger_cursor', $1)`,
-				fmt.Sprintf("%d", tc.latestIngested))
-			require.NoError(t, err)
-			_, err = dbConnectionPool.Exec(ctx,
-				`INSERT INTO ingest_store (key, value) VALUES ('oldest_ledger_cursor', $1)`,
-				fmt.Sprintf("%d", tc.latestIngested))
-			require.NoError(t, err)
-
-			mockMetricsService := metrics.NewMockMetricsService()
-			mockMetricsService.On("RegisterPoolMetrics", "ledger_indexer", mock.Anything).Return()
-			mockMetricsService.On("RegisterPoolMetrics", "backfill", mock.Anything).Return()
-			mockMetricsService.On("ObserveDBQueryDuration", mock.Anything, mock.Anything, mock.Anything).Return().Maybe()
-			mockMetricsService.On("IncDBQuery", mock.Anything, mock.Anything).Return().Maybe()
-			defer mockMetricsService.AssertExpectations(t)
-
-			models, err := data.NewModels(dbConnectionPool, mockMetricsService)
-			require.NoError(t, err)
-
-			mockBackendFactory := func(ctx context.Context) (ledgerbackend.LedgerBackend, error) {
-				return nil, fmt.Errorf("mock backend factory error")
-			}
-
-			svc, err := NewIngestService(IngestServiceConfig{
-				IngestionMode:          IngestionModeBackfill,
-				Models:                 models,
-				LatestLedgerCursorName: "latest_ledger_cursor",
-				OldestLedgerCursorName: "oldest_ledger_cursor",
-				AppTracker:             &apptracker.MockAppTracker{},
-				RPCService:             &RPCServiceMock{},
-				LedgerBackend:          &LedgerBackendMock{},
-				LedgerBackendFactory:   mockBackendFactory,
-				ChannelAccountStore:    &store.ChannelAccountStoreMock{},
-				MetricsService:         mockMetricsService,
-				GetLedgersLimit:        defaultGetLedgersLimit,
-				Network:                network.TestNetworkPassphrase,
-				NetworkPassphrase:      network.TestNetworkPassphrase,
-				Archive:                &HistoryArchiveMock{},
-				SkipTxMeta:             false,
-				SkipTxEnvelope:         false,
-				BackfillBatchSize:      100,
-			})
-			require.NoError(t, err)
-
-			err = svc.startCatchup(ctx, tc.startLedger, tc.endLedger)
-			if tc.expectValidationError {
-				require.Error(t, err)
-				assert.Contains(t, err.Error(), tc.errorContains)
-			} else {
-				// Catchup mode returns error when batches fail
-				require.Error(t, err)
-				assert.Contains(t, err.Error(), "optimized catchup failed")
-				assert.NotContains(t, err.Error(), "catchup must start from ledger")
-			}
-		})
-	}
-}
-
-// ==================== Test Helper Functions ====================
-
-// setupDBCursors sets up ingest_store cursors for testing
-func setupDBCursors(t *testing.T, ctx context.Context, pool *pgxpool.Pool, latestLedger, oldestLedger uint32) {
-	_, err := pool.Exec(ctx, `DELETE FROM ingest_store`)
-	require.NoError(t, err)
-	if latestLedger > 0 {
-		_, err = pool.Exec(ctx, `INSERT INTO ingest_store (key, value) VALUES ('latest_ledger_cursor', $1)`, fmt.Sprintf("%d", latestLedger))
-		require.NoError(t, err)
-	}
-	if oldestLedger > 0 {
-		_, err = pool.Exec(ctx, `INSERT INTO ingest_store (key, value) VALUES ('oldest_ledger_cursor', $1)`, fmt.Sprintf("%d", oldestLedger))
-		require.NoError(t, err)
-	}
-}
-
-// ptrUint32 returns a pointer to the given uint32 value
-func ptrUint32(v uint32) *uint32 { return &v }
-
-// createTestTransaction creates a transaction with required fields for testing.
-func createTestTransaction(hash string, toID int64) types.Transaction {
-	now := time.Now()
-	envelope := "test_envelope_xdr"
-	meta := "test_meta_xdr"
-	return types.Transaction{
-		Hash:            types.HashBytea(hash),
-		ToID:            toID,
-		EnvelopeXDR:     &envelope,
-		FeeCharged:      100,
-		ResultCode:      "TransactionResultCodeTxSuccess",
-		MetaXDR:         &meta,
-		LedgerNumber:    1000,
-		LedgerCreatedAt: now,
-		IngestedAt:      now,
-	}
-}
-
-// createTestOperation creates an operation with required fields for testing.
-func createTestOperation(id int64) types.Operation {
-	now := time.Now()
-	return types.Operation{
-		ID:              id,
-		OperationType:   types.OperationTypePayment,
-		OperationXDR:    types.XDRBytea([]byte("test_operation_xdr")),
-		LedgerNumber:    1000,
-		LedgerCreatedAt: now,
-		IngestedAt:      now,
-	}
-}
-
-// createTestStateChange creates a state change with required fields for testing.
-func createTestStateChange(toID int64, accountID string, opID int64) types.StateChange {
-	now := time.Now()
-	reason := types.StateChangeReasonCredit
-	return types.StateChange{
-		ToID:                toID,
-		StateChangeOrder:    1,
-		StateChangeCategory: types.StateChangeCategoryBalance,
-		StateChangeReason:   &reason,
-		AccountID:           types.AddressBytea(accountID),
-		OperationID:         opID,
-		LedgerNumber:        1000,
-		LedgerCreatedAt:     now,
-		IngestedAt:          now,
-	}
-}
-
-// ==================== New Tests ====================
 
 func Test_analyzeBatchResults(t *testing.T) {
 	ctx := context.Background()
@@ -884,87 +414,6 @@ func Test_ingestService_setupBatchBackend(t *testing.T) {
 	}
 }
 
-func Test_ingestService_updateOldestCursor(t *testing.T) {
-	dbt := dbtest.Open(t)
-	defer dbt.Close()
-	ctx := context.Background()
-	dbConnectionPool, err := db.OpenDBConnectionPool(ctx, dbt.DSN)
-	require.NoError(t, err)
-	defer dbConnectionPool.Close()
-
-	testCases := []struct {
-		name          string
-		initialCursor uint32
-		updateTo      uint32
-		wantCursor    uint32 // Expected cursor after update (UpdateMin keeps lower)
-	}{
-		{
-			name:          "updates_to_lower_value",
-			initialCursor: 100,
-			updateTo:      50,
-			wantCursor:    50,
-		},
-		{
-			name:          "keeps_existing_lower_value",
-			initialCursor: 50,
-			updateTo:      100,
-			wantCursor:    50, // UpdateMin keeps the lower value
-		},
-		{
-			name:          "same_value_no_change",
-			initialCursor: 100,
-			updateTo:      100,
-			wantCursor:    100,
-		},
-	}
-
-	for _, tc := range testCases {
-		t.Run(tc.name, func(t *testing.T) {
-			// Clean up and set initial cursor
-			setupDBCursors(t, ctx, dbConnectionPool, 200, tc.initialCursor)
-
-			mockMetricsService := metrics.NewMockMetricsService()
-			mockMetricsService.On("RegisterPoolMetrics", "ledger_indexer", mock.Anything).Return()
-			mockMetricsService.On("RegisterPoolMetrics", "backfill", mock.Anything).Return()
-			mockMetricsService.On("ObserveDBQueryDuration", mock.Anything, mock.Anything, mock.Anything).Return().Maybe()
-			mockMetricsService.On("IncDBQuery", mock.Anything, mock.Anything).Return().Maybe()
-			mockMetricsService.On("IncDBTransaction", mock.Anything).Return().Maybe()
-			mockMetricsService.On("ObserveDBTransactionDuration", mock.Anything, mock.Anything).Return().Maybe()
-			defer mockMetricsService.AssertExpectations(t)
-
-			models, err := data.NewModels(dbConnectionPool, mockMetricsService)
-			require.NoError(t, err)
-
-			mockRPCService := &RPCServiceMock{}
-			mockRPCService.On("NetworkPassphrase").Return(network.TestNetworkPassphrase).Maybe()
-
-			svc, err := NewIngestService(IngestServiceConfig{
-				IngestionMode:          IngestionModeBackfill,
-				Models:                 models,
-				LatestLedgerCursorName: "latest_ledger_cursor",
-				OldestLedgerCursorName: "oldest_ledger_cursor",
-				AppTracker:             &apptracker.MockAppTracker{},
-				RPCService:             mockRPCService,
-				LedgerBackend:          &LedgerBackendMock{},
-				MetricsService:         mockMetricsService,
-				GetLedgersLimit:        defaultGetLedgersLimit,
-				Network:                network.TestNetworkPassphrase,
-				NetworkPassphrase:      network.TestNetworkPassphrase,
-				Archive:                &HistoryArchiveMock{},
-			})
-			require.NoError(t, err)
-
-			err = svc.updateOldestCursor(ctx, tc.updateTo)
-			require.NoError(t, err)
-
-			// Verify the cursor value
-			cursor, err := models.IngestStore.Get(ctx, "oldest_ledger_cursor")
-			require.NoError(t, err)
-			assert.Equal(t, tc.wantCursor, cursor)
-		})
-	}
-}
-
 func Test_ingestService_initializeCursors(t *testing.T) {
 	dbt := dbtest.Open(t)
 	defer dbt.Close()
@@ -1101,780 +550,6 @@ func Test_ingestService_Run(t *testing.T) {
 			assert.Contains(t, err.Error(), tc.wantErrContains)
 		})
 	}
-}
-
-func Test_ingestService_flushHistoricalBatch(t *testing.T) {
-	dbt := dbtest.Open(t)
-	defer dbt.Close()
-	ctx := context.Background()
-	dbConnectionPool, err := db.OpenDBConnectionPool(ctx, dbt.DSN)
-	require.NoError(t, err)
-	defer dbConnectionPool.Close()
-
-	testCases := []struct {
-		name                 string
-		setupBuffer          func() *indexer.IndexerBuffer
-		updateCursorTo       *uint32
-		initialCursor        uint32
-		wantCursor           uint32
-		wantTxCount          int
-		wantOpCount          int
-		wantStateChangeCount int
-		txHashes             []string // For verification queries
-	}{
-		{
-			name:                 "flush_empty_buffer_no_cursor_update",
-			setupBuffer:          func() *indexer.IndexerBuffer { return indexer.NewIndexerBuffer() },
-			updateCursorTo:       nil,
-			initialCursor:        100,
-			wantCursor:           100,
-			wantTxCount:          0,
-			wantOpCount:          0,
-			wantStateChangeCount: 0,
-			txHashes:             []string{},
-		},
-		{
-			name: "flush_with_data_inserts_to_database",
-			setupBuffer: func() *indexer.IndexerBuffer {
-				buf := indexer.NewIndexerBuffer()
-				tx1 := createTestTransaction(flushTxHash1, 1)
-				tx2 := createTestTransaction(flushTxHash2, 2)
-				op1 := createTestOperation(200)
-				op2 := createTestOperation(201)
-				sc1 := createTestStateChange(1, testAddr1, 200)
-				sc2 := createTestStateChange(2, testAddr2, 201)
-
-				buf.PushTransaction(testAddr1, tx1)
-				buf.PushTransaction(testAddr2, tx2)
-				buf.PushOperation(testAddr1, op1, tx1)
-				buf.PushOperation(testAddr2, op2, tx2)
-				buf.PushStateChange(tx1, op1, sc1)
-				buf.PushStateChange(tx2, op2, sc2)
-				return buf
-			},
-			updateCursorTo:       nil,
-			initialCursor:        100,
-			wantCursor:           100,
-			wantTxCount:          2,
-			wantOpCount:          2,
-			wantStateChangeCount: 2,
-			txHashes:             []string{flushTxHash1, flushTxHash2},
-		},
-		{
-			name: "flush_with_cursor_update_to_lower_value",
-			setupBuffer: func() *indexer.IndexerBuffer {
-				buf := indexer.NewIndexerBuffer()
-				tx1 := createTestTransaction(flushTxHash3, 3)
-				buf.PushTransaction(testAddr1, tx1)
-				return buf
-			},
-			updateCursorTo:       ptrUint32(50),
-			initialCursor:        100,
-			wantCursor:           50,
-			wantTxCount:          1,
-			wantOpCount:          0,
-			wantStateChangeCount: 0,
-			txHashes:             []string{flushTxHash3},
-		},
-		{
-			name: "flush_with_cursor_update_to_higher_value_keeps_existing",
-			setupBuffer: func() *indexer.IndexerBuffer {
-				buf := indexer.NewIndexerBuffer()
-				tx1 := createTestTransaction(flushTxHash4, 4)
-				buf.PushTransaction(testAddr1, tx1)
-				return buf
-			},
-			updateCursorTo:       ptrUint32(150),
-			initialCursor:        100,
-			wantCursor:           100, // UpdateMin keeps lower
-			wantTxCount:          1,
-			wantOpCount:          0,
-			wantStateChangeCount: 0,
-			txHashes:             []string{flushTxHash4},
-		},
-	}
-
-	for _, tc := range testCases {
-		t.Run(tc.name, func(t *testing.T) {
-			// Clean up test data from previous runs (using HashBytea for BYTEA column)
-			for _, hash := range []string{flushTxHash1, flushTxHash2, flushTxHash3, flushTxHash4} {
-				_, err = dbConnectionPool.Exec(ctx, `DELETE FROM state_changes WHERE to_id IN (SELECT to_id FROM transactions WHERE hash = $1)`, types.HashBytea(hash))
-				require.NoError(t, err)
-				_, err = dbConnectionPool.Exec(ctx, `DELETE FROM transactions WHERE hash = $1`, types.HashBytea(hash))
-				require.NoError(t, err)
-			}
-			// Also clean up any orphan operations
-			_, err = dbConnectionPool.Exec(ctx, `TRUNCATE operations, operations_accounts CASCADE`)
-			require.NoError(t, err)
-
-			// Set up initial cursor
-			setupDBCursors(t, ctx, dbConnectionPool, 200, tc.initialCursor)
-
-			mockMetricsService := metrics.NewMockMetricsService()
-			mockMetricsService.On("RegisterPoolMetrics", "ledger_indexer", mock.Anything).Return()
-			mockMetricsService.On("RegisterPoolMetrics", "backfill", mock.Anything).Return()
-			mockMetricsService.On("ObserveDBQueryDuration", mock.Anything, mock.Anything, mock.Anything).Return().Maybe()
-			mockMetricsService.On("IncDBQuery", mock.Anything, mock.Anything).Return().Maybe()
-			mockMetricsService.On("IncDBTransaction", mock.Anything).Return().Maybe()
-			mockMetricsService.On("ObserveDBTransactionDuration", mock.Anything, mock.Anything).Return().Maybe()
-			mockMetricsService.On("ObserveDBBatchSize", mock.Anything, mock.Anything, mock.Anything).Return().Maybe()
-			mockMetricsService.On("IncStateChanges", mock.Anything, mock.Anything, mock.Anything).Return().Maybe()
-			defer mockMetricsService.AssertExpectations(t)
-
-			models, err := data.NewModels(dbConnectionPool, mockMetricsService)
-			require.NoError(t, err)
-
-			mockRPCService := &RPCServiceMock{}
-			mockRPCService.On("NetworkPassphrase").Return(network.TestNetworkPassphrase).Maybe()
-
-			mockChAccStore := &store.ChannelAccountStoreMock{}
-
-			svc, err := NewIngestService(IngestServiceConfig{
-				IngestionMode:          IngestionModeBackfill,
-				Models:                 models,
-				LatestLedgerCursorName: "latest_ledger_cursor",
-				OldestLedgerCursorName: "oldest_ledger_cursor",
-				AppTracker:             &apptracker.MockAppTracker{},
-				RPCService:             mockRPCService,
-				LedgerBackend:          &LedgerBackendMock{},
-				ChannelAccountStore:    mockChAccStore,
-				MetricsService:         mockMetricsService,
-				GetLedgersLimit:        defaultGetLedgersLimit,
-				Network:                network.TestNetworkPassphrase,
-				NetworkPassphrase:      network.TestNetworkPassphrase,
-				Archive:                &HistoryArchiveMock{},
-			})
-			require.NoError(t, err)
-
-			buffer := tc.setupBuffer()
-
-			// Call flushHistoricalBatch
-			err = svc.flushHistoricalBatch(ctx, buffer, tc.updateCursorTo)
-			require.NoError(t, err)
-
-			// Verify the cursor value
-			cursor, err := models.IngestStore.Get(ctx, "oldest_ledger_cursor")
-			require.NoError(t, err)
-			assert.Equal(t, tc.wantCursor, cursor)
-
-			// Verify transaction count in database
-			if len(tc.txHashes) > 0 {
-				hashBytes := make([][]byte, len(tc.txHashes))
-				for i, h := range tc.txHashes {
-					val, err := types.HashBytea(h).Value()
-					require.NoError(t, err)
-					hashBytes[i] = val.([]byte)
-				}
-				var txCount int
-				err = dbConnectionPool.QueryRow(ctx,
-					`SELECT COUNT(*) FROM transactions WHERE hash = ANY($1)`,
-					hashBytes).Scan(&txCount)
-				require.NoError(t, err)
-				assert.Equal(t, tc.wantTxCount, txCount, "transaction count mismatch")
-			}
-
-			// Verify operation count in database
-			if tc.wantOpCount > 0 {
-				var opCount int
-				err = dbConnectionPool.QueryRow(ctx,
-					`SELECT COUNT(*) FROM operations`).Scan(&opCount)
-				require.NoError(t, err)
-				assert.Equal(t, tc.wantOpCount, opCount, "operation count mismatch")
-			}
-
-			// Verify state change count in database
-			if tc.wantStateChangeCount > 0 {
-				scHashBytes := make([][]byte, len(tc.txHashes))
-				for i, h := range tc.txHashes {
-					val, err := types.HashBytea(h).Value()
-					require.NoError(t, err)
-					scHashBytes[i] = val.([]byte)
-				}
-				var scCount int
-				err = dbConnectionPool.QueryRow(ctx,
-					`SELECT COUNT(*) FROM state_changes WHERE to_id IN (SELECT to_id FROM transactions WHERE hash = ANY($1))`,
-					scHashBytes).Scan(&scCount)
-				require.NoError(t, err)
-				assert.Equal(t, tc.wantStateChangeCount, scCount, "state change count mismatch")
-			}
-
-			// Historical backfill should never unlock channel accounts
-			mockChAccStore.AssertNotCalled(t, "UnassignTxAndUnlockChannelAccounts")
-		})
-	}
-}
-
-// ==================== Backfill Failure Scenario Tests ====================
-
-// Test_ingestService_processBackfillBatchesParallel_PartialFailure verifies that when one
-// batch fails during parallel processing, other batches still complete successfully.
-func Test_ingestService_processBackfillBatchesParallel_PartialFailure(t *testing.T) {
-	dbt := dbtest.Open(t)
-	defer dbt.Close()
-	ctx := context.Background()
-	dbConnectionPool, err := db.OpenDBConnectionPool(ctx, dbt.DSN)
-	require.NoError(t, err)
-	defer dbConnectionPool.Close()
-
-	testCases := []struct {
-		name            string
-		batches         []BackfillBatch
-		failBatchIndex  int // which batch index should fail (0-based)
-		wantFailedCount int
-	}{
-		{
-			name: "middle_batch_fails",
-			batches: []BackfillBatch{
-				{StartLedger: 100, EndLedger: 109},
-				{StartLedger: 110, EndLedger: 119}, // This one fails
-				{StartLedger: 120, EndLedger: 129},
-			},
-			failBatchIndex:  1,
-			wantFailedCount: 1,
-		},
-		{
-			name: "first_batch_fails",
-			batches: []BackfillBatch{
-				{StartLedger: 100, EndLedger: 109}, // This one fails
-				{StartLedger: 110, EndLedger: 119},
-				{StartLedger: 120, EndLedger: 129},
-			},
-			failBatchIndex:  0,
-			wantFailedCount: 1,
-		},
-		{
-			name: "last_batch_fails",
-			batches: []BackfillBatch{
-				{StartLedger: 100, EndLedger: 109},
-				{StartLedger: 110, EndLedger: 119},
-				{StartLedger: 120, EndLedger: 129}, // This one fails
-			},
-			failBatchIndex:  2,
-			wantFailedCount: 1,
-		},
-	}
-
-	for _, tc := range testCases {
-		t.Run(tc.name, func(t *testing.T) {
-			mockMetricsService := metrics.NewMockMetricsService()
-			mockMetricsService.On("RegisterPoolMetrics", "ledger_indexer", mock.Anything).Return()
-			mockMetricsService.On("RegisterPoolMetrics", "backfill", mock.Anything).Return()
-			mockMetricsService.On("SetOldestLedgerIngested", mock.Anything).Return().Maybe()
-			mockMetricsService.On("ObserveDBQueryDuration", mock.Anything, mock.Anything, mock.Anything).Return().Maybe()
-			mockMetricsService.On("IncDBQuery", mock.Anything, mock.Anything).Return().Maybe()
-			mockMetricsService.On("IncDBTransaction", mock.Anything).Return().Maybe()
-			mockMetricsService.On("ObserveDBTransactionDuration", mock.Anything, mock.Anything).Return().Maybe()
-			mockMetricsService.On("ObserveDBBatchSize", mock.Anything, mock.Anything, mock.Anything).Return().Maybe()
-			mockMetricsService.On("ObserveIngestionParticipantsCount", mock.Anything).Return().Maybe()
-			mockMetricsService.On("IncStateChanges", mock.Anything, mock.Anything, mock.Anything).Return().Maybe()
-			defer mockMetricsService.AssertExpectations(t)
-
-			models, modelsErr := data.NewModels(dbConnectionPool, mockMetricsService)
-			require.NoError(t, modelsErr)
-
-			mockRPCService := &RPCServiceMock{}
-			mockRPCService.On("NetworkPassphrase").Return(network.TestNetworkPassphrase).Maybe()
-
-			failBatch := tc.batches[tc.failBatchIndex]
-
-			// Factory that returns a backend that fails PrepareRange for the specified batch
-			factory := func(ctx context.Context) (ledgerbackend.LedgerBackend, error) {
-				mockBackend := &LedgerBackendMock{}
-				// Use MatchedBy to check if this is the failing batch
-				mockBackend.On("PrepareRange", mock.Anything, mock.MatchedBy(func(r ledgerbackend.Range) bool {
-					return r.From() == failBatch.StartLedger
-				})).Return(fmt.Errorf("simulated failure for batch starting at %d", failBatch.StartLedger))
-				// All other batches succeed - return error to prevent processing
-				// This avoids nil pointer issues with empty LedgerCloseMeta
-				mockBackend.On("PrepareRange", mock.Anything, mock.Anything).Return(nil).Maybe()
-				// Return proper empty ledger meta to avoid nil pointer issues
-				mockBackend.On("GetLedger", mock.Anything, mock.Anything).Return(xdr.LedgerCloseMeta{
-					V: 0,
-					V0: &xdr.LedgerCloseMetaV0{
-						LedgerHeader: xdr.LedgerHeaderHistoryEntry{
-							Header: xdr.LedgerHeader{
-								LedgerSeq: xdr.Uint32(100),
-							},
-						},
-					},
-				}, nil).Maybe()
-				mockBackend.On("Close").Return(nil).Maybe()
-				return mockBackend, nil
-			}
-
-			svc, svcErr := NewIngestService(IngestServiceConfig{
-				IngestionMode:          IngestionModeBackfill,
-				Models:                 models,
-				LatestLedgerCursorName: "latest_ledger_cursor",
-				OldestLedgerCursorName: "oldest_ledger_cursor",
-				AppTracker:             &apptracker.MockAppTracker{},
-				RPCService:             mockRPCService,
-				LedgerBackend:          &LedgerBackendMock{},
-				LedgerBackendFactory:   factory,
-				MetricsService:         mockMetricsService,
-				GetLedgersLimit:        defaultGetLedgersLimit,
-				Network:                network.TestNetworkPassphrase,
-				NetworkPassphrase:      network.TestNetworkPassphrase,
-				Archive:                &HistoryArchiveMock{},
-				BackfillBatchSize:      10,
-			})
-			require.NoError(t, svcErr)
-
-			results := svc.processBackfillBatchesParallel(ctx, tc.batches, svc.processLedgersInBatchHistorical, nil)
-
-			// Verify results
-			require.Len(t, results, len(tc.batches))
-
-			// Count actual failures
-			actualFailed := 0
-			for i, result := range results {
-				if result.Error != nil {
-					actualFailed++
-					assert.Equal(t, tc.failBatchIndex, i, "expected batch at index %d to fail", tc.failBatchIndex)
-					assert.Contains(t, result.Error.Error(), "preparing backend range")
-				}
-			}
-			assert.Equal(t, tc.wantFailedCount, actualFailed)
-
-			// Verify analyzeBatchResults returns correct count
-			numFailed := analyzeBatchResults(ctx, results)
-			assert.Equal(t, tc.wantFailedCount, numFailed)
-		})
-	}
-}
-
-// Test_ingestService_startBackfilling_HistoricalMode_PartialFailure_CursorUpdate verifies
-// that the oldest_ingest_ledger cursor is updated to the minimum of successful batches
-// when some batches fail during historical backfill.
-func Test_ingestService_startBackfilling_HistoricalMode_PartialFailure_CursorUpdate(t *testing.T) {
-	dbt := dbtest.Open(t)
-	defer dbt.Close()
-	ctx := context.Background()
-	dbConnectionPool, err := db.OpenDBConnectionPool(ctx, dbt.DSN)
-	require.NoError(t, err)
-	defer dbConnectionPool.Close()
-
-	testCases := []struct {
-		name            string
-		initialOldest   uint32
-		initialLatest   uint32
-		startLedger     uint32
-		endLedger       uint32
-		batchSize       uint32
-		failBatchStart  uint32 // start ledger of batch to fail
-		wantFinalOldest uint32
-		wantError       bool
-		wantErrContains string
-	}{
-		{
-			name:            "middle_batch_fails_cursor_updates_to_minimum",
-			initialOldest:   100,
-			initialLatest:   100,
-			startLedger:     50,
-			endLedger:       99,
-			batchSize:       10,
-			failBatchStart:  70, // Batch [70-79] fails
-			wantFinalOldest: 50, // Minimum of successful batches (50, 60, 80, 90)
-			wantError:       false,
-		},
-		{
-			name:            "first_batch_fails_cursor_updates_to_second_batch",
-			initialOldest:   100,
-			initialLatest:   100,
-			startLedger:     50,
-			endLedger:       79,
-			batchSize:       10,
-			failBatchStart:  50, // Batch [50-59] fails
-			wantFinalOldest: 60, // Next successful batch
-			wantError:       false,
-		},
-	}
-
-	for _, tc := range testCases {
-		t.Run(tc.name, func(t *testing.T) {
-			// Clean up database
-			_, err = dbConnectionPool.Exec(ctx, `DELETE FROM transactions`)
-			require.NoError(t, err)
-			_, err = dbConnectionPool.Exec(ctx, `DELETE FROM operations`)
-			require.NoError(t, err)
-			_, err = dbConnectionPool.Exec(ctx, `DELETE FROM state_changes`)
-			require.NoError(t, err)
-
-			// Set up initial cursors
-			setupDBCursors(t, ctx, dbConnectionPool, tc.initialLatest, tc.initialOldest)
-
-			// Insert anchor transaction so GetOldestLedger() returns initialOldest
-			_, err = dbConnectionPool.Exec(ctx,
-				`INSERT INTO transactions (hash, to_id, envelope_xdr, fee_charged, result_code, meta_xdr, ledger_number, ledger_created_at)
-				VALUES ('anchor_hash', 1, 'env', 100, 'TransactionResultCodeTxSuccess', 'meta', $1, NOW())`,
-				tc.initialOldest)
-			require.NoError(t, err)
-
-			mockMetricsService := metrics.NewMockMetricsService()
-			mockMetricsService.On("RegisterPoolMetrics", "ledger_indexer", mock.Anything).Return()
-			mockMetricsService.On("RegisterPoolMetrics", "backfill", mock.Anything).Return()
-			mockMetricsService.On("SetOldestLedgerIngested", mock.Anything).Return().Maybe()
-			mockMetricsService.On("ObserveDBQueryDuration", mock.Anything, mock.Anything, mock.Anything).Return().Maybe()
-			mockMetricsService.On("IncDBQuery", mock.Anything, mock.Anything).Return().Maybe()
-			mockMetricsService.On("IncDBTransaction", mock.Anything).Return().Maybe()
-			mockMetricsService.On("ObserveDBTransactionDuration", mock.Anything, mock.Anything).Return().Maybe()
-			mockMetricsService.On("ObserveDBBatchSize", mock.Anything, mock.Anything, mock.Anything).Return().Maybe()
-			mockMetricsService.On("IncStateChanges", mock.Anything, mock.Anything, mock.Anything).Return().Maybe()
-			mockMetricsService.On("ObserveIngestionParticipantsCount", mock.Anything).Return().Maybe()
-			defer mockMetricsService.AssertExpectations(t)
-
-			models, modelsErr := data.NewModels(dbConnectionPool, mockMetricsService)
-			require.NoError(t, modelsErr)
-
-			mockRPCService := &RPCServiceMock{}
-			mockRPCService.On("NetworkPassphrase").Return(network.TestNetworkPassphrase).Maybe()
-
-			// Factory that fails for the specified batch
-			factory := func(ctx context.Context) (ledgerbackend.LedgerBackend, error) {
-				mockBackend := &LedgerBackendMock{}
-				mockBackend.On("PrepareRange", mock.Anything, mock.MatchedBy(func(r ledgerbackend.Range) bool {
-					return r.From() == tc.failBatchStart
-				})).Return(fmt.Errorf("simulated network failure"))
-				mockBackend.On("PrepareRange", mock.Anything, mock.Anything).Return(nil).Maybe()
-				// For successful batches, return empty ledgers (no transactions)
-				mockBackend.On("GetLedger", mock.Anything, mock.Anything).Return(xdr.LedgerCloseMeta{
-					V: 0,
-					V0: &xdr.LedgerCloseMetaV0{
-						LedgerHeader: xdr.LedgerHeaderHistoryEntry{
-							Header: xdr.LedgerHeader{
-								LedgerSeq: xdr.Uint32(50),
-							},
-						},
-					},
-				}, nil).Maybe()
-				mockBackend.On("Close").Return(nil).Maybe()
-				return mockBackend, nil
-			}
-
-			svc, svcErr := NewIngestService(IngestServiceConfig{
-				IngestionMode:          IngestionModeBackfill,
-				Models:                 models,
-				LatestLedgerCursorName: "latest_ledger_cursor",
-				OldestLedgerCursorName: "oldest_ledger_cursor",
-				AppTracker:             &apptracker.MockAppTracker{},
-				RPCService:             mockRPCService,
-				LedgerBackend:          &LedgerBackendMock{},
-				LedgerBackendFactory:   factory,
-				MetricsService:         mockMetricsService,
-				GetLedgersLimit:        defaultGetLedgersLimit,
-				Network:                network.TestNetworkPassphrase,
-				NetworkPassphrase:      network.TestNetworkPassphrase,
-				Archive:                &HistoryArchiveMock{},
-				BackfillBatchSize:      int(tc.batchSize),
-			})
-			require.NoError(t, svcErr)
-
-			// Run backfilling
-			backfillErr := svc.startHistoricalBackfill(ctx, tc.startLedger, tc.endLedger)
-
-			if tc.wantError {
-				require.Error(t, backfillErr)
-				if tc.wantErrContains != "" {
-					assert.Contains(t, backfillErr.Error(), tc.wantErrContains)
-				}
-			} else {
-				// Historical mode should not return error even with partial failures
-				require.NoError(t, backfillErr)
-			}
-
-			// Verify cursor was updated correctly using IngestStore.Get
-			finalOldest, getErr := models.IngestStore.Get(ctx, "oldest_ledger_cursor")
-			require.NoError(t, getErr)
-			assert.Equal(t, tc.wantFinalOldest, finalOldest,
-				"oldest cursor should be updated to minimum of successful batches")
-		})
-	}
-}
-
-// Test_ingestService_processBackfillBatches_PartialFailure_OnlySuccessfulBatchPersisted verifies
-// that when one batch fails and another succeeds:
-// - The failed batch does not persist any data
-// - The successful batch persists its transactions
-// - Proper error handling for both cases
-func Test_ingestService_processBackfillBatches_PartialFailure_OnlySuccessfulBatchPersisted(t *testing.T) {
-	dbt := dbtest.Open(t)
-	defer dbt.Close()
-	ctx := context.Background()
-	dbConnectionPool, err := db.OpenDBConnectionPool(ctx, dbt.DSN)
-	require.NoError(t, err)
-	defer dbConnectionPool.Close()
-
-	// Clean up database
-	_, err = dbConnectionPool.Exec(ctx, `DELETE FROM transactions`)
-	require.NoError(t, err)
-	_, err = dbConnectionPool.Exec(ctx, `DELETE FROM operations`)
-	require.NoError(t, err)
-	_, err = dbConnectionPool.Exec(ctx, `DELETE FROM state_changes`)
-	require.NoError(t, err)
-
-	// Set up initial cursors
-	setupDBCursors(t, ctx, dbConnectionPool, 200, 200)
-
-	mockMetricsService := metrics.NewMockMetricsService()
-	mockMetricsService.On("RegisterPoolMetrics", "ledger_indexer", mock.Anything).Return()
-	mockMetricsService.On("RegisterPoolMetrics", "backfill", mock.Anything).Return()
-	mockMetricsService.On("SetOldestLedgerIngested", mock.Anything).Return().Maybe()
-	mockMetricsService.On("ObserveDBQueryDuration", mock.Anything, mock.Anything, mock.Anything).Return().Maybe()
-	mockMetricsService.On("IncDBQuery", mock.Anything, mock.Anything).Return().Maybe()
-	mockMetricsService.On("IncDBTransaction", mock.Anything).Return().Maybe()
-	mockMetricsService.On("ObserveDBTransactionDuration", mock.Anything, mock.Anything).Return().Maybe()
-	mockMetricsService.On("ObserveDBBatchSize", mock.Anything, mock.Anything, mock.Anything).Return().Maybe()
-	mockMetricsService.On("ObserveIngestionParticipantsCount", mock.Anything).Return().Maybe()
-	mockMetricsService.On("IncStateChanges", mock.Anything, mock.Anything, mock.Anything).Return().Maybe()
-	mockMetricsService.On("ObserveStateChangeProcessingDuration", mock.Anything, mock.Anything).Return().Maybe()
-	mockMetricsService.On("IncDBQueryError", mock.Anything, mock.Anything, mock.Anything).Return().Maybe()
-	mockMetricsService.On("ObserveIngestionDuration", mock.Anything, mock.Anything).Return().Maybe()
-	defer mockMetricsService.AssertExpectations(t)
-
-	models, modelsErr := data.NewModels(dbConnectionPool, mockMetricsService)
-	require.NoError(t, modelsErr)
-
-	mockRPCService := &RPCServiceMock{}
-	mockRPCService.On("NetworkPassphrase").Return(network.TestNetworkPassphrase).Maybe()
-
-	// Parse ledger metadata with a transaction for the successful batch
-	var metaWithTx xdr.LedgerCloseMeta
-	err = xdr.SafeUnmarshalBase64(ledgerMetadataWith1Tx, &metaWithTx)
-	require.NoError(t, err)
-
-	// Two batches: first fails, second succeeds
-	// Use single-ledger batches to avoid duplicate key issues with test fixtures
-	batches := []BackfillBatch{
-		{StartLedger: 100, EndLedger: 100}, // Will fail
-		{StartLedger: 110, EndLedger: 110}, // Will succeed - single ledger to avoid duplicate tx hash
-	}
-
-	// Factory that returns a mock backend configured based on the batch range
-	factory := func(ctx context.Context) (ledgerbackend.LedgerBackend, error) {
-		mockBackend := &LedgerBackendMock{}
-		// Fail batch 1 (100-109) at PrepareRange
-		mockBackend.On("PrepareRange", mock.Anything, mock.MatchedBy(func(r ledgerbackend.Range) bool {
-			return r.From() == 100
-		})).Return(fmt.Errorf("ledger range unavailable"))
-		// Succeed batch 2 (110-119)
-		mockBackend.On("PrepareRange", mock.Anything, mock.Anything).Return(nil).Maybe()
-		mockBackend.On("GetLedger", mock.Anything, mock.Anything).Return(metaWithTx, nil).Maybe()
-		mockBackend.On("Close").Return(nil).Maybe()
-		return mockBackend, nil
-	}
-
-	svc, svcErr := NewIngestService(IngestServiceConfig{
-		IngestionMode:             IngestionModeBackfill,
-		Models:                    models,
-		LatestLedgerCursorName:    "latest_ledger_cursor",
-		OldestLedgerCursorName:    "oldest_ledger_cursor",
-		AppTracker:                &apptracker.MockAppTracker{},
-		RPCService:                mockRPCService,
-		LedgerBackend:             &LedgerBackendMock{},
-		LedgerBackendFactory:      factory,
-		MetricsService:            mockMetricsService,
-		GetLedgersLimit:           defaultGetLedgersLimit,
-		Network:                   network.TestNetworkPassphrase,
-		NetworkPassphrase:         network.TestNetworkPassphrase,
-		Archive:                   &HistoryArchiveMock{},
-		BackfillBatchSize:         10,
-		BackfillDBInsertBatchSize: 50,
-	})
-	require.NoError(t, svcErr)
-
-	// Process both batches in parallel
-	results := svc.processBackfillBatchesParallel(ctx, batches, svc.processLedgersInBatchHistorical, nil)
-
-	// Verify we got results for both batches
-	require.Len(t, results, 2)
-
-	// Verify batch 1 (100-109) failed
-	require.Error(t, results[0].Error, "batch 1 should have failed")
-	assert.Contains(t, results[0].Error.Error(), "preparing backend range")
-	assert.Contains(t, results[0].Error.Error(), "ledger range unavailable")
-
-	// Verify batch 2 (110-119) succeeded
-	require.NoError(t, results[1].Error, "batch 2 should have succeeded")
-
-	// Verify no transactions were persisted for failed batch (ledger 100)
-	var failedBatchTxCount int
-	err = dbConnectionPool.QueryRow(ctx,
-		`SELECT COUNT(*) FROM transactions WHERE ledger_number BETWEEN $1 AND $2`,
-		100, 109).Scan(&failedBatchTxCount)
-	require.NoError(t, err)
-	assert.Equal(t, 0, failedBatchTxCount, "no transactions should be persisted for failed batch")
-
-	// Verify transactions were persisted for successful batch
-	// The ledgerMetadataWith1Tx fixture has a fixed ledger sequence of 4478
-	// so we query for that ledger number (the XDR metadata determines the stored ledger)
-	var successBatchTxCount int
-	err = dbConnectionPool.QueryRow(ctx,
-		`SELECT COUNT(*) FROM transactions WHERE ledger_number = $1`,
-		4478).Scan(&successBatchTxCount)
-	require.NoError(t, err)
-	assert.Equal(t, 1, successBatchTxCount, "1 transaction should be persisted for successful batch")
-}
-
-// Test_ingestService_startBackfilling_CatchupMode_PartialFailure_ReturnsError verifies
-// that catchup mode returns an error when any batch fails.
-func Test_ingestService_startBackfilling_CatchupMode_PartialFailure_ReturnsError(t *testing.T) {
-	dbt := dbtest.Open(t)
-	defer dbt.Close()
-	ctx := context.Background()
-	dbConnectionPool, err := db.OpenDBConnectionPool(ctx, dbt.DSN)
-	require.NoError(t, err)
-	defer dbConnectionPool.Close()
-
-	// Clean up database
-	_, err = dbConnectionPool.Exec(ctx, `DELETE FROM transactions`)
-	require.NoError(t, err)
-
-	// Set up initial cursors: latest = 99, so catchup starts at 100
-	setupDBCursors(t, ctx, dbConnectionPool, 99, 50)
-
-	mockMetricsService := metrics.NewMockMetricsService()
-	mockMetricsService.On("RegisterPoolMetrics", "ledger_indexer", mock.Anything).Return()
-	mockMetricsService.On("RegisterPoolMetrics", "backfill", mock.Anything).Return()
-	mockMetricsService.On("ObserveDBQueryDuration", mock.Anything, mock.Anything, mock.Anything).Return().Maybe()
-	mockMetricsService.On("IncDBQuery", mock.Anything, mock.Anything).Return().Maybe()
-	mockMetricsService.On("IncDBTransaction", mock.Anything).Return().Maybe()
-	mockMetricsService.On("ObserveDBTransactionDuration", mock.Anything, mock.Anything).Return().Maybe()
-	mockMetricsService.On("ObserveDBBatchSize", mock.Anything, mock.Anything, mock.Anything).Return().Maybe()
-	mockMetricsService.On("ObserveIngestionParticipantsCount", mock.Anything).Return().Maybe()
-	mockMetricsService.On("IncStateChanges", mock.Anything, mock.Anything, mock.Anything).Return().Maybe()
-	defer mockMetricsService.AssertExpectations(t)
-
-	models, modelsErr := data.NewModels(dbConnectionPool, mockMetricsService)
-	require.NoError(t, modelsErr)
-
-	mockRPCService := &RPCServiceMock{}
-	mockRPCService.On("NetworkPassphrase").Return(network.TestNetworkPassphrase).Maybe()
-
-	// Factory that fails for the middle batch
-	factory := func(ctx context.Context) (ledgerbackend.LedgerBackend, error) {
-		mockBackend := &LedgerBackendMock{}
-		// Fail for batch starting at 110
-		mockBackend.On("PrepareRange", mock.Anything, mock.MatchedBy(func(r ledgerbackend.Range) bool {
-			return r.From() == 110
-		})).Return(fmt.Errorf("simulated catchup failure"))
-		mockBackend.On("PrepareRange", mock.Anything, mock.Anything).Return(nil).Maybe()
-		mockBackend.On("GetLedger", mock.Anything, mock.Anything).Return(xdr.LedgerCloseMeta{
-			V: 0,
-			V0: &xdr.LedgerCloseMetaV0{
-				LedgerHeader: xdr.LedgerHeaderHistoryEntry{
-					Header: xdr.LedgerHeader{
-						LedgerSeq: xdr.Uint32(100),
-					},
-				},
-			},
-		}, nil).Maybe()
-		mockBackend.On("Close").Return(nil).Maybe()
-		return mockBackend, nil
-	}
-
-	svc, svcErr := NewIngestService(IngestServiceConfig{
-		IngestionMode:          IngestionModeBackfill,
-		Models:                 models,
-		LatestLedgerCursorName: "latest_ledger_cursor",
-		OldestLedgerCursorName: "oldest_ledger_cursor",
-		AppTracker:             &apptracker.MockAppTracker{},
-		RPCService:             mockRPCService,
-		LedgerBackend:          &LedgerBackendMock{},
-		LedgerBackendFactory:   factory,
-		MetricsService:         mockMetricsService,
-		GetLedgersLimit:        defaultGetLedgersLimit,
-		Network:                network.TestNetworkPassphrase,
-		NetworkPassphrase:      network.TestNetworkPassphrase,
-		Archive:                &HistoryArchiveMock{},
-		BackfillBatchSize:      10,
-	})
-	require.NoError(t, svcErr)
-
-	// Run catchup backfilling: starts at latest+1 = 100, ends at 129 (3 batches)
-	backfillErr := svc.startCatchup(ctx, 100, 129)
-
-	// Verify error is returned for catchup mode
-	require.Error(t, backfillErr)
-	assert.Contains(t, backfillErr.Error(), "optimized catchup failed")
-	assert.Contains(t, backfillErr.Error(), "batches failed")
-
-	// Verify latest cursor was NOT updated
-	finalLatest, getErr := models.IngestStore.Get(ctx, "latest_ledger_cursor")
-	require.NoError(t, getErr)
-	assert.Equal(t, uint32(99), finalLatest, "latest cursor should NOT be updated when catchup fails")
-}
-
-// Test_ingestService_startBackfilling_HistoricalMode_AllBatchesFail_CursorUnchanged
-// verifies that the cursor remains unchanged when all batches fail during historical backfill.
-func Test_ingestService_startBackfilling_HistoricalMode_AllBatchesFail_CursorUnchanged(t *testing.T) {
-	dbt := dbtest.Open(t)
-	defer dbt.Close()
-	ctx := context.Background()
-	dbConnectionPool, err := db.OpenDBConnectionPool(ctx, dbt.DSN)
-	require.NoError(t, err)
-	defer dbConnectionPool.Close()
-
-	// Clean up database
-	_, err = dbConnectionPool.Exec(ctx, `DELETE FROM transactions`)
-	require.NoError(t, err)
-
-	// Set up initial cursors
-	initialOldest := uint32(100)
-	initialLatest := uint32(100)
-	setupDBCursors(t, ctx, dbConnectionPool, initialLatest, initialOldest)
-
-	mockMetricsService := metrics.NewMockMetricsService()
-	mockMetricsService.On("RegisterPoolMetrics", "ledger_indexer", mock.Anything).Return()
-	mockMetricsService.On("RegisterPoolMetrics", "backfill", mock.Anything).Return()
-	mockMetricsService.On("ObserveDBQueryDuration", mock.Anything, mock.Anything, mock.Anything).Return().Maybe()
-	mockMetricsService.On("IncDBQuery", mock.Anything, mock.Anything).Return().Maybe()
-	defer mockMetricsService.AssertExpectations(t)
-
-	models, modelsErr := data.NewModels(dbConnectionPool, mockMetricsService)
-	require.NoError(t, modelsErr)
-
-	mockRPCService := &RPCServiceMock{}
-	mockRPCService.On("NetworkPassphrase").Return(network.TestNetworkPassphrase).Maybe()
-
-	// Factory that always fails
-	factory := func(ctx context.Context) (ledgerbackend.LedgerBackend, error) {
-		return nil, fmt.Errorf("all backends unavailable")
-	}
-
-	svc, svcErr := NewIngestService(IngestServiceConfig{
-		IngestionMode:          IngestionModeBackfill,
-		Models:                 models,
-		LatestLedgerCursorName: "latest_ledger_cursor",
-		OldestLedgerCursorName: "oldest_ledger_cursor",
-		AppTracker:             &apptracker.MockAppTracker{},
-		RPCService:             mockRPCService,
-		LedgerBackend:          &LedgerBackendMock{},
-		LedgerBackendFactory:   factory,
-		MetricsService:         mockMetricsService,
-		GetLedgersLimit:        defaultGetLedgersLimit,
-		Network:                network.TestNetworkPassphrase,
-		NetworkPassphrase:      network.TestNetworkPassphrase,
-		Archive:                &HistoryArchiveMock{},
-		BackfillBatchSize:      10,
-	})
-	require.NoError(t, svcErr)
-
-	// Run backfilling with all batches failing
-	backfillErr := svc.startHistoricalBackfill(ctx, 50, 99)
-
-	// Historical mode should NOT return error even when all batches fail
-	require.NoError(t, backfillErr, "historical mode should not return error even when all batches fail")
-
-	// Verify cursor remains unchanged
-	finalOldest, getErr := models.IngestStore.Get(ctx, "oldest_ledger_cursor")
-	require.NoError(t, getErr)
-	assert.Equal(t, initialOldest, finalOldest,
-		"oldest cursor should remain unchanged when all batches fail")
-
-	finalLatest, getErr := models.IngestStore.Get(ctx, "latest_ledger_cursor")
-	require.NoError(t, getErr)
-	assert.Equal(t, initialLatest, finalLatest,
-		"latest cursor should remain unchanged when all batches fail")
 }
 
 // Test_ingestProcessedDataWithRetry tests the ingestProcessedDataWithRetry function covering success, failure, and retry scenarios.
@@ -2153,820 +828,4 @@ func Test_ingestProcessedDataWithRetry(t *testing.T) {
 
 		mockTokenIngestionService.AssertExpectations(t)
 	})
-}
-
-// Test_ingestService_processBatchChanges tests the processBatchChanges method which processes
-// aggregated batch changes after all parallel batches complete.
-// NOTE: The implementation now:
-// 1. Calls models.TrustlineAsset.BatchInsert for trustline assets (real DB operation)
-// 2. Calls contractMetadataService.FetchMetadata for new contracts (mocked)
-// 3. Calls models.Contract.BatchInsert for contracts (real DB operation)
-// 4. Calls tokenIngestionService.ProcessTokenChanges(ctx, dbTx, trustlineChanges, contractChanges) (mocked)
-func Test_ingestService_processBatchChanges(t *testing.T) {
-	dbt := dbtest.Open(t)
-	defer dbt.Close()
-	ctx := context.Background()
-	dbConnectionPool, err := db.OpenDBConnectionPool(ctx, dbt.DSN)
-	require.NoError(t, err)
-	defer dbConnectionPool.Close()
-
-	testCases := []struct {
-		name                 string
-		trustlineChanges     map[indexer.TrustlineChangeKey]types.TrustlineChange
-		contractChanges      []types.ContractChange
-		uniqueAssets         map[uuid.UUID]data.TrustlineAsset
-		uniqueContractTokens map[string]types.ContractType
-		setupMocks           func(t *testing.T, tokenIngestionService *TokenIngestionServiceMock, contractMetadataSvc *ContractMetadataServiceMock)
-		wantErr              bool
-		wantErrContains      string
-		verifySortedChanges  func(t *testing.T, tokenIngestionService *TokenIngestionServiceMock)
-	}{
-		{
-			name:                 "empty_data_calls_ProcessTokenChanges",
-			trustlineChanges:     map[indexer.TrustlineChangeKey]types.TrustlineChange{},
-			contractChanges:      []types.ContractChange{},
-			uniqueAssets:         map[uuid.UUID]data.TrustlineAsset{},
-			uniqueContractTokens: map[string]types.ContractType{},
-			setupMocks: func(t *testing.T, tokenIngestionService *TokenIngestionServiceMock, contractMetadataSvc *ContractMetadataServiceMock) {
-				// ProcessTokenChanges is called with empty map and slice
-				tokenIngestionService.On("ProcessTokenChanges", mock.Anything, mock.Anything, mock.MatchedBy(func(changes map[indexer.TrustlineChangeKey]types.TrustlineChange) bool {
-					return len(changes) == 0
-				}), []types.ContractChange{}, mock.MatchedBy(func(m map[string]types.AccountChange) bool { return true }), mock.Anything).Return(nil)
-			},
-			wantErr: false,
-		},
-		{
-			// Data is now pre-deduplicated by caller
-			name: "passes_through_deduplicated_trustline_changes",
-			trustlineChanges: map[indexer.TrustlineChangeKey]types.TrustlineChange{
-				{AccountID: "GA1", TrustlineID: data.DeterministicAssetID("USD", "GA1")}: {AccountID: "GA1", Asset: "USD:GA1", OperationID: toid.New(100, 0, 10).ToInt64(), LedgerNumber: 100},
-				{AccountID: "GA2", TrustlineID: data.DeterministicAssetID("EUR", "GA2")}: {AccountID: "GA2", Asset: "EUR:GA2", OperationID: toid.New(101, 0, 1).ToInt64(), LedgerNumber: 101},
-				{AccountID: "GA3", TrustlineID: data.DeterministicAssetID("GBP", "GA3")}: {AccountID: "GA3", Asset: "GBP:GA3", OperationID: toid.New(100, 0, 5).ToInt64(), LedgerNumber: 100},
-			},
-			contractChanges:      []types.ContractChange{},
-			uniqueAssets:         map[uuid.UUID]data.TrustlineAsset{},
-			uniqueContractTokens: map[string]types.ContractType{},
-			setupMocks: func(t *testing.T, tokenIngestionService *TokenIngestionServiceMock, contractMetadataSvc *ContractMetadataServiceMock) {
-				// Verify all 3 changes are in the map (different keys)
-				tokenIngestionService.On("ProcessTokenChanges", mock.Anything, mock.Anything, mock.MatchedBy(func(changes map[indexer.TrustlineChangeKey]types.TrustlineChange) bool {
-					return len(changes) == 3
-				}), []types.ContractChange{}, mock.MatchedBy(func(m map[string]types.AccountChange) bool { return true }), mock.Anything).Return(nil)
-			},
-			wantErr: false,
-		},
-		{
-			// Contract changes are passed through (not deduplicated here)
-			name:             "passes_contract_changes_through",
-			trustlineChanges: map[indexer.TrustlineChangeKey]types.TrustlineChange{},
-			contractChanges: []types.ContractChange{
-				{AccountID: "GA1", ContractID: "C1", OperationID: toid.New(200, 0, 20).ToInt64(), LedgerNumber: 200, ContractType: types.ContractTypeUnknown},
-				{AccountID: "GA2", ContractID: "C2", OperationID: toid.New(200, 0, 5).ToInt64(), LedgerNumber: 200, ContractType: types.ContractTypeUnknown},
-				{AccountID: "GA3", ContractID: "C3", OperationID: toid.New(201, 0, 1).ToInt64(), LedgerNumber: 201, ContractType: types.ContractTypeUnknown},
-			},
-			uniqueAssets:         map[uuid.UUID]data.TrustlineAsset{},
-			uniqueContractTokens: map[string]types.ContractType{},
-			setupMocks: func(t *testing.T, tokenIngestionService *TokenIngestionServiceMock, contractMetadataSvc *ContractMetadataServiceMock) {
-				// All 3 contract changes are passed through
-				tokenIngestionService.On("ProcessTokenChanges", mock.Anything, mock.Anything, mock.MatchedBy(func(changes map[indexer.TrustlineChangeKey]types.TrustlineChange) bool {
-					return len(changes) == 0
-				}), mock.MatchedBy(func(changes []types.ContractChange) bool {
-					return len(changes) == 3
-				}), mock.MatchedBy(func(m map[string]types.AccountChange) bool { return true }), mock.Anything).Return(nil)
-			},
-			wantErr: false,
-		},
-		{
-			name: "inserts_trustline_assets_and_calls_ProcessTokenChanges",
-			trustlineChanges: map[indexer.TrustlineChangeKey]types.TrustlineChange{
-				{AccountID: "GA1", TrustlineID: data.DeterministicAssetID("USDC", "GA5ZSEJYB37JRC5AVCIA5MOP4RHTM335X2KGX3IHOJAPP5RE34K4KZVN")}: {AccountID: "GA1", Asset: "USDC:GA5ZSEJYB37JRC5AVCIA5MOP4RHTM335X2KGX3IHOJAPP5RE34K4KZVN", OperationID: 1, LedgerNumber: 100, Operation: types.TrustlineOpAdd},
-			},
-			contractChanges: []types.ContractChange{},
-			uniqueAssets: map[uuid.UUID]data.TrustlineAsset{
-				data.DeterministicAssetID("USDC", "GA5ZSEJYB37JRC5AVCIA5MOP4RHTM335X2KGX3IHOJAPP5RE34K4KZVN"): {ID: data.DeterministicAssetID("USDC", "GA5ZSEJYB37JRC5AVCIA5MOP4RHTM335X2KGX3IHOJAPP5RE34K4KZVN"), Code: "USDC", Issuer: "GA5ZSEJYB37JRC5AVCIA5MOP4RHTM335X2KGX3IHOJAPP5RE34K4KZVN"},
-			},
-			uniqueContractTokens: map[string]types.ContractType{},
-			setupMocks: func(t *testing.T, tokenIngestionService *TokenIngestionServiceMock, contractMetadataSvc *ContractMetadataServiceMock) {
-				// ProcessTokenChanges is called after trustline assets are inserted to DB
-				tokenIngestionService.On("ProcessTokenChanges", mock.Anything, mock.Anything, mock.MatchedBy(func(changes map[indexer.TrustlineChangeKey]types.TrustlineChange) bool {
-					if len(changes) != 1 {
-						return false
-					}
-					for _, change := range changes {
-						return change.Asset == "USDC:GA5ZSEJYB37JRC5AVCIA5MOP4RHTM335X2KGX3IHOJAPP5RE34K4KZVN"
-					}
-					return false
-				}), []types.ContractChange{}, mock.MatchedBy(func(m map[string]types.AccountChange) bool { return true }), mock.Anything).Return(nil)
-			},
-			wantErr: false,
-		},
-		{
-			// Note: Unknown contracts don't get inserted into contracts table but are still passed to ProcessTokenChanges
-			name:                 "processes_unknown_contract_changes",
-			trustlineChanges:     map[indexer.TrustlineChangeKey]types.TrustlineChange{},
-			uniqueAssets:         map[uuid.UUID]data.TrustlineAsset{},
-			uniqueContractTokens: map[string]types.ContractType{},
-			contractChanges: []types.ContractChange{
-				{AccountID: "GA1", ContractID: "CDLZFC3SYJYDZT7K67VZ75HPJVIEUVNIXF47ZG2FB2RMQQVU2HHGCYSC", OperationID: toid.New(100, 0, 1).ToInt64(), LedgerNumber: 100, ContractType: types.ContractTypeUnknown},
-			},
-			setupMocks: func(t *testing.T, tokenIngestionService *TokenIngestionServiceMock, contractMetadataSvc *ContractMetadataServiceMock) {
-				// Unknown contracts don't trigger FetchMetadata but are passed to ProcessTokenChanges
-				tokenIngestionService.On("ProcessTokenChanges", mock.Anything, mock.Anything, mock.MatchedBy(func(changes map[indexer.TrustlineChangeKey]types.TrustlineChange) bool {
-					return len(changes) == 0
-				}), mock.MatchedBy(func(changes []types.ContractChange) bool {
-					return len(changes) == 1 && changes[0].ContractType == types.ContractTypeUnknown
-				}), mock.MatchedBy(func(m map[string]types.AccountChange) bool { return true }), mock.Anything).Return(nil)
-			},
-			wantErr: false,
-		},
-		{
-			name: "calls_ProcessTokenChanges_with_trustline_and_contract_changes",
-			trustlineChanges: map[indexer.TrustlineChangeKey]types.TrustlineChange{
-				{AccountID: "GA1", TrustlineID: data.DeterministicAssetID("USDC", "GA")}: {AccountID: "GA1", Asset: "USDC:GA", OperationID: toid.New(100, 0, 1).ToInt64(), LedgerNumber: 100},
-			},
-			contractChanges: []types.ContractChange{
-				{AccountID: "GA2", ContractID: "C1", OperationID: toid.New(100, 0, 2).ToInt64(), LedgerNumber: 100, ContractType: types.ContractTypeUnknown},
-			},
-			uniqueAssets:         map[uuid.UUID]data.TrustlineAsset{},
-			uniqueContractTokens: map[string]types.ContractType{},
-			setupMocks: func(t *testing.T, tokenIngestionService *TokenIngestionServiceMock, contractMetadataSvc *ContractMetadataServiceMock) {
-				// ProcessTokenChanges receives both trustline and contract changes
-				tokenIngestionService.On("ProcessTokenChanges", mock.Anything, mock.Anything, mock.MatchedBy(func(changes map[indexer.TrustlineChangeKey]types.TrustlineChange) bool {
-					if len(changes) != 1 {
-						return false
-					}
-					for _, change := range changes {
-						return change.Asset == "USDC:GA"
-					}
-					return false
-				}), mock.MatchedBy(func(changes []types.ContractChange) bool {
-					return len(changes) == 1 && changes[0].ContractID == "C1"
-				}), mock.MatchedBy(func(m map[string]types.AccountChange) bool { return true }), mock.Anything).Return(nil)
-			},
-			wantErr: false,
-		},
-		{
-			name: "propagates_ProcessTokenChanges_error",
-			trustlineChanges: map[indexer.TrustlineChangeKey]types.TrustlineChange{
-				{AccountID: "GA1", TrustlineID: data.DeterministicAssetID("USDC", "GA")}: {AccountID: "GA1", Asset: "USDC:GA", OperationID: toid.New(100, 0, 1).ToInt64(), LedgerNumber: 100},
-			},
-			contractChanges:      []types.ContractChange{},
-			uniqueAssets:         map[uuid.UUID]data.TrustlineAsset{},
-			uniqueContractTokens: map[string]types.ContractType{},
-			setupMocks: func(t *testing.T, tokenIngestionService *TokenIngestionServiceMock, contractMetadataSvc *ContractMetadataServiceMock) {
-				// ProcessTokenChanges receives the actual trustline changes
-				tokenIngestionService.On("ProcessTokenChanges", mock.Anything, mock.Anything, mock.MatchedBy(func(changes map[indexer.TrustlineChangeKey]types.TrustlineChange) bool {
-					return len(changes) == 1
-				}), []types.ContractChange{}, mock.MatchedBy(func(m map[string]types.AccountChange) bool { return true }), mock.Anything).Return(fmt.Errorf("db error"))
-			},
-			wantErr:         true,
-			wantErrContains: "processing token changes",
-		},
-	}
-
-	for _, tc := range testCases {
-		t.Run(tc.name, func(t *testing.T) {
-			mockMetricsService := metrics.NewMockMetricsService()
-			mockMetricsService.On("RegisterPoolMetrics", "ledger_indexer", mock.Anything).Return()
-			mockMetricsService.On("RegisterPoolMetrics", "backfill", mock.Anything).Return()
-			mockMetricsService.On("ObserveDBQueryDuration", mock.Anything, mock.Anything, mock.Anything).Return().Maybe()
-			mockMetricsService.On("IncDBQuery", mock.Anything, mock.Anything).Return().Maybe()
-			mockMetricsService.On("IncDBTransaction", mock.Anything).Return().Maybe()
-			mockMetricsService.On("ObserveDBTransactionDuration", mock.Anything, mock.Anything).Return().Maybe()
-			defer mockMetricsService.AssertExpectations(t)
-
-			models, err := data.NewModels(dbConnectionPool, mockMetricsService)
-			require.NoError(t, err)
-
-			mockRPCService := &RPCServiceMock{}
-			mockRPCService.On("NetworkPassphrase").Return(network.TestNetworkPassphrase).Maybe()
-
-			mockTokenIngestionService := NewTokenIngestionServiceMock(t)
-			mockContractMetadataSvc := NewContractMetadataServiceMock(t)
-
-			tc.setupMocks(t, mockTokenIngestionService, mockContractMetadataSvc)
-
-			svc, err := NewIngestService(IngestServiceConfig{
-				IngestionMode:           IngestionModeBackfill,
-				Models:                  models,
-				LatestLedgerCursorName:  "latest_ledger_cursor",
-				OldestLedgerCursorName:  "oldest_ledger_cursor",
-				AppTracker:              &apptracker.MockAppTracker{},
-				RPCService:              mockRPCService,
-				LedgerBackend:           &LedgerBackendMock{},
-				TokenIngestionService:   mockTokenIngestionService,
-				ContractMetadataService: mockContractMetadataSvc,
-				MetricsService:          mockMetricsService,
-				GetLedgersLimit:         defaultGetLedgersLimit,
-				Network:                 network.TestNetworkPassphrase,
-				NetworkPassphrase:       network.TestNetworkPassphrase,
-				Archive:                 &HistoryArchiveMock{},
-			})
-			require.NoError(t, err)
-
-			err = db.RunInTransaction(ctx, models.DB, func(dbTx pgx.Tx) error {
-				return svc.processBatchChanges(ctx, dbTx, tc.trustlineChanges, tc.contractChanges, make(map[string]types.AccountChange), make(map[indexer.SACBalanceChangeKey]types.SACBalanceChange), tc.uniqueAssets, tc.uniqueContractTokens, make(map[string]*data.Contract))
-			})
-
-			if tc.wantErr {
-				require.Error(t, err)
-				if tc.wantErrContains != "" {
-					assert.Contains(t, err.Error(), tc.wantErrContains)
-				}
-			} else {
-				require.NoError(t, err)
-			}
-
-			if tc.verifySortedChanges != nil {
-				tc.verifySortedChanges(t, mockTokenIngestionService)
-			}
-		})
-	}
-}
-
-// Test_ingestService_flushCatchupBatch tests that batch changes are collected
-// when flushCatchupBatch is called.
-func Test_ingestService_flushCatchupBatch(t *testing.T) {
-	dbt := dbtest.Open(t)
-	defer dbt.Close()
-	ctx := context.Background()
-	dbConnectionPool, err := db.OpenDBConnectionPool(ctx, dbt.DSN)
-	require.NoError(t, err)
-	defer dbConnectionPool.Close()
-
-	testCases := []struct {
-		name                      string
-		setupBuffer               func() *indexer.IndexerBuffer
-		batchChanges              *BatchChanges
-		wantTrustlineChangesCount int
-		wantContractChanges       []types.ContractChange
-	}{
-		{
-			name: "collects_trustline_changes_when_batchChanges_provided",
-			setupBuffer: func() *indexer.IndexerBuffer {
-				buf := indexer.NewIndexerBuffer()
-				tx1 := createTestTransaction(catchupTxHash1, 1)
-				buf.PushTransaction(testAddr1, tx1)
-				buf.PushTrustlineChange(types.TrustlineChange{
-					AccountID:    testAddr1,
-					Asset:        "USDC:GA5ZSEJYB37JRC5AVCIA5MOP4RHTM335X2KGX3IHOJAPP5RE34K4KZVN",
-					OperationID:  100,
-					LedgerNumber: 1000,
-					Operation:    types.TrustlineOpAdd,
-				})
-				return buf
-			},
-			batchChanges:              &BatchChanges{TrustlineChangesByKey: make(map[indexer.TrustlineChangeKey]types.TrustlineChange), UniqueTrustlineAssets: make(map[uuid.UUID]data.TrustlineAsset), UniqueContractTokensByID: make(map[string]types.ContractType)},
-			wantTrustlineChangesCount: 1,
-			wantContractChanges:       nil,
-		},
-		{
-			name: "collects_contract_changes_when_batchChanges_provided",
-			setupBuffer: func() *indexer.IndexerBuffer {
-				buf := indexer.NewIndexerBuffer()
-				tx1 := createTestTransaction(catchupTxHash2, 2)
-				buf.PushTransaction(testAddr2, tx1)
-				buf.PushContractChange(types.ContractChange{
-					AccountID:    testAddr2,
-					ContractID:   "CCONTRACTID",
-					OperationID:  101,
-					LedgerNumber: 1001,
-					ContractType: types.ContractTypeSAC,
-				})
-				return buf
-			},
-			batchChanges:              &BatchChanges{TrustlineChangesByKey: make(map[indexer.TrustlineChangeKey]types.TrustlineChange), UniqueTrustlineAssets: make(map[uuid.UUID]data.TrustlineAsset), UniqueContractTokensByID: make(map[string]types.ContractType)},
-			wantTrustlineChangesCount: 0,
-			wantContractChanges: []types.ContractChange{
-				{
-					AccountID:    testAddr2,
-					ContractID:   "CCONTRACTID",
-					OperationID:  101,
-					LedgerNumber: 1001,
-					ContractType: types.ContractTypeSAC,
-				},
-			},
-		},
-		{
-			name: "accumulates_across_multiple_flushes",
-			setupBuffer: func() *indexer.IndexerBuffer {
-				buf := indexer.NewIndexerBuffer()
-				tx1 := createTestTransaction(catchupTxHash6, 6)
-				buf.PushTransaction(testAddr1, tx1)
-				buf.PushTrustlineChange(types.TrustlineChange{
-					AccountID:    testAddr1,
-					Asset:        "GBP:GA5ZSEJYB37JRC5AVCIA5MOP4RHTM335X2KGX3IHOJAPP5RE34K4KZVN",
-					OperationID:  103,
-					LedgerNumber: 1003,
-					Operation:    types.TrustlineOpAdd,
-				})
-				return buf
-			},
-			// Pre-populate batchChanges to simulate accumulation from previous flush
-			batchChanges: &BatchChanges{
-				TrustlineChangesByKey: map[indexer.TrustlineChangeKey]types.TrustlineChange{
-					{AccountID: "GPREV", TrustlineID: data.DeterministicAssetID("PREV", "GA5ZSEJYB37JRC5AVCIA5MOP4RHTM335X2KGX3IHOJAPP5RE34K4KZVN")}: {AccountID: "GPREV", Asset: "PREV:GA5ZSEJYB37JRC5AVCIA5MOP4RHTM335X2KGX3IHOJAPP5RE34K4KZVN", OperationID: 50, LedgerNumber: 999, Operation: types.TrustlineOpAdd},
-				},
-				UniqueTrustlineAssets:    make(map[uuid.UUID]data.TrustlineAsset),
-				UniqueContractTokensByID: make(map[string]types.ContractType),
-			},
-			wantTrustlineChangesCount: 2, // Pre-existing + new change
-			wantContractChanges:       nil,
-		},
-	}
-
-	for _, tc := range testCases {
-		t.Run(tc.name, func(t *testing.T) {
-			// Clean up test data from previous runs (using HashBytea for BYTEA column)
-			for _, hash := range []string{catchupTxHash1, catchupTxHash2, catchupTxHash3, catchupTxHash4, catchupTxHash5, catchupTxHash6, prevTxHash} {
-				_, err = dbConnectionPool.Exec(ctx, `DELETE FROM state_changes WHERE to_id IN (SELECT to_id FROM transactions WHERE hash = $1)`, types.HashBytea(hash))
-				require.NoError(t, err)
-				_, err = dbConnectionPool.Exec(ctx, `DELETE FROM transactions WHERE hash = $1`, types.HashBytea(hash))
-				require.NoError(t, err)
-			}
-			// Also clean up any orphan operations
-			_, err = dbConnectionPool.Exec(ctx, `TRUNCATE operations, operations_accounts CASCADE`)
-			require.NoError(t, err)
-
-			// Set up cursors
-			setupDBCursors(t, ctx, dbConnectionPool, 200, 100)
-
-			mockMetricsService := metrics.NewMockMetricsService()
-			mockMetricsService.On("RegisterPoolMetrics", "ledger_indexer", mock.Anything).Return()
-			mockMetricsService.On("RegisterPoolMetrics", "backfill", mock.Anything).Return()
-			mockMetricsService.On("ObserveDBQueryDuration", mock.Anything, mock.Anything, mock.Anything).Return().Maybe()
-			mockMetricsService.On("IncDBQuery", mock.Anything, mock.Anything).Return().Maybe()
-			mockMetricsService.On("IncDBTransaction", mock.Anything).Return().Maybe()
-			mockMetricsService.On("ObserveDBTransactionDuration", mock.Anything, mock.Anything).Return().Maybe()
-			mockMetricsService.On("ObserveDBBatchSize", mock.Anything, mock.Anything, mock.Anything).Return().Maybe()
-			mockMetricsService.On("IncStateChanges", mock.Anything, mock.Anything, mock.Anything).Return().Maybe()
-			defer mockMetricsService.AssertExpectations(t)
-
-			models, err := data.NewModels(dbConnectionPool, mockMetricsService)
-			require.NoError(t, err)
-
-			mockRPCService := &RPCServiceMock{}
-			mockRPCService.On("NetworkPassphrase").Return(network.TestNetworkPassphrase).Maybe()
-
-			svc, err := NewIngestService(IngestServiceConfig{
-				IngestionMode:          IngestionModeBackfill,
-				Models:                 models,
-				LatestLedgerCursorName: "latest_ledger_cursor",
-				OldestLedgerCursorName: "oldest_ledger_cursor",
-				AppTracker:             &apptracker.MockAppTracker{},
-				RPCService:             mockRPCService,
-				LedgerBackend:          &LedgerBackendMock{},
-				MetricsService:         mockMetricsService,
-				GetLedgersLimit:        defaultGetLedgersLimit,
-				Network:                network.TestNetworkPassphrase,
-				NetworkPassphrase:      network.TestNetworkPassphrase,
-				Archive:                &HistoryArchiveMock{},
-			})
-			require.NoError(t, err)
-
-			buffer := tc.setupBuffer()
-
-			err = svc.flushCatchupBatch(ctx, buffer, tc.batchChanges)
-			require.NoError(t, err)
-
-			// Verify collected token changes match expected values
-			require.Len(t, tc.batchChanges.TrustlineChangesByKey, tc.wantTrustlineChangesCount, "trustline changes count mismatch")
-
-			require.Len(t, tc.batchChanges.ContractChanges, len(tc.wantContractChanges), "contract changes count mismatch")
-			for i, want := range tc.wantContractChanges {
-				got := tc.batchChanges.ContractChanges[i]
-				assert.Equal(t, want.AccountID, got.AccountID, "ContractChange[%d].AccountID mismatch", i)
-				assert.Equal(t, want.ContractID, got.ContractID, "ContractChange[%d].ContractID mismatch", i)
-				assert.Equal(t, want.OperationID, got.OperationID, "ContractChange[%d].OperationID mismatch", i)
-				assert.Equal(t, want.LedgerNumber, got.LedgerNumber, "ContractChange[%d].LedgerNumber mismatch", i)
-				assert.Equal(t, want.ContractType, got.ContractType, "ContractChange[%d].ContractType mismatch", i)
-			}
-		})
-	}
-}
-
-// Test_ingestService_processLedgersInBatchCatchup tests that processLedgersInBatchCatchup
-// returns non-nil BatchChanges.
-func Test_ingestService_processLedgersInBatchCatchup(t *testing.T) {
-	dbt := dbtest.Open(t)
-	defer dbt.Close()
-	ctx := context.Background()
-	dbConnectionPool, err := db.OpenDBConnectionPool(ctx, dbt.DSN)
-	require.NoError(t, err)
-	defer dbConnectionPool.Close()
-
-	var ledgerMeta xdr.LedgerCloseMeta
-	err = xdr.SafeUnmarshalBase64(ledgerMetadataWith0Tx, &ledgerMeta)
-	require.NoError(t, err)
-
-	setupDBCursors(t, ctx, dbConnectionPool, 200, 100)
-
-	mockMetricsService := metrics.NewMockMetricsService()
-	mockMetricsService.On("RegisterPoolMetrics", "ledger_indexer", mock.Anything).Return()
-	mockMetricsService.On("RegisterPoolMetrics", "backfill", mock.Anything).Return()
-	mockMetricsService.On("ObserveDBQueryDuration", mock.Anything, mock.Anything, mock.Anything).Return().Maybe()
-	mockMetricsService.On("IncDBQuery", mock.Anything, mock.Anything).Return().Maybe()
-	mockMetricsService.On("IncDBTransaction", mock.Anything).Return().Maybe()
-	mockMetricsService.On("ObserveDBTransactionDuration", mock.Anything, mock.Anything).Return().Maybe()
-	mockMetricsService.On("ObserveDBBatchSize", mock.Anything, mock.Anything, mock.Anything).Return().Maybe()
-	mockMetricsService.On("IncStateChanges", mock.Anything, mock.Anything, mock.Anything).Return().Maybe()
-	mockMetricsService.On("ObserveIngestionParticipantsCount", mock.Anything).Return().Maybe()
-	defer mockMetricsService.AssertExpectations(t)
-
-	models, err := data.NewModels(dbConnectionPool, mockMetricsService)
-	require.NoError(t, err)
-
-	mockLedgerBackend := &LedgerBackendMock{}
-	mockLedgerBackend.On("GetLedger", mock.Anything, uint32(4599)).Return(ledgerMeta, nil)
-
-	mockChAccStore := &store.ChannelAccountStoreMock{}
-	mockChAccStore.On("UnassignTxAndUnlockChannelAccounts", mock.Anything, mock.Anything).Return(int64(0), nil).Maybe()
-
-	svc, err := NewIngestService(IngestServiceConfig{
-		IngestionMode:          IngestionModeBackfill,
-		Models:                 models,
-		LatestLedgerCursorName: "latest_ledger_cursor",
-		OldestLedgerCursorName: "oldest_ledger_cursor",
-		AppTracker:             &apptracker.MockAppTracker{},
-		RPCService:             &RPCServiceMock{},
-		LedgerBackend:          mockLedgerBackend,
-		ChannelAccountStore:    mockChAccStore,
-		MetricsService:         mockMetricsService,
-		GetLedgersLimit:        defaultGetLedgersLimit,
-		Network:                network.TestNetworkPassphrase,
-		NetworkPassphrase:      network.TestNetworkPassphrase,
-		Archive:                &HistoryArchiveMock{},
-	})
-	require.NoError(t, err)
-
-	batch := BackfillBatch{StartLedger: 4599, EndLedger: 4599}
-	result := svc.processLedgersInBatchCatchup(ctx, mockLedgerBackend, batch)
-
-	require.NoError(t, result.Error)
-	assert.Equal(t, 1, result.LedgersCount)
-	assert.NotNil(t, result.BatchChanges, "expected non-nil batch changes for catchup mode")
-}
-
-// Test_ingestService_processLedgersInBatchHistorical tests that processLedgersInBatchHistorical
-// returns nil BatchChanges.
-func Test_ingestService_processLedgersInBatchHistorical(t *testing.T) {
-	dbt := dbtest.Open(t)
-	defer dbt.Close()
-	ctx := context.Background()
-	dbConnectionPool, err := db.OpenDBConnectionPool(ctx, dbt.DSN)
-	require.NoError(t, err)
-	defer dbConnectionPool.Close()
-
-	var ledgerMeta xdr.LedgerCloseMeta
-	err = xdr.SafeUnmarshalBase64(ledgerMetadataWith0Tx, &ledgerMeta)
-	require.NoError(t, err)
-
-	setupDBCursors(t, ctx, dbConnectionPool, 200, 100)
-
-	mockMetricsService := metrics.NewMockMetricsService()
-	mockMetricsService.On("RegisterPoolMetrics", "ledger_indexer", mock.Anything).Return()
-	mockMetricsService.On("RegisterPoolMetrics", "backfill", mock.Anything).Return()
-	mockMetricsService.On("SetOldestLedgerIngested", mock.Anything).Return().Maybe()
-	mockMetricsService.On("ObserveDBQueryDuration", mock.Anything, mock.Anything, mock.Anything).Return().Maybe()
-	mockMetricsService.On("IncDBQuery", mock.Anything, mock.Anything).Return().Maybe()
-	mockMetricsService.On("IncDBTransaction", mock.Anything).Return().Maybe()
-	mockMetricsService.On("ObserveDBTransactionDuration", mock.Anything, mock.Anything).Return().Maybe()
-	mockMetricsService.On("ObserveDBBatchSize", mock.Anything, mock.Anything, mock.Anything).Return().Maybe()
-	mockMetricsService.On("IncStateChanges", mock.Anything, mock.Anything, mock.Anything).Return().Maybe()
-	mockMetricsService.On("ObserveIngestionParticipantsCount", mock.Anything).Return().Maybe()
-	defer mockMetricsService.AssertExpectations(t)
-
-	models, err := data.NewModels(dbConnectionPool, mockMetricsService)
-	require.NoError(t, err)
-
-	mockLedgerBackend := &LedgerBackendMock{}
-	mockLedgerBackend.On("GetLedger", mock.Anything, uint32(4599)).Return(ledgerMeta, nil)
-
-	mockChAccStore := &store.ChannelAccountStoreMock{}
-	mockChAccStore.On("UnassignTxAndUnlockChannelAccounts", mock.Anything, mock.Anything).Return(int64(0), nil).Maybe()
-
-	svc, err := NewIngestService(IngestServiceConfig{
-		IngestionMode:          IngestionModeBackfill,
-		Models:                 models,
-		LatestLedgerCursorName: "latest_ledger_cursor",
-		OldestLedgerCursorName: "oldest_ledger_cursor",
-		AppTracker:             &apptracker.MockAppTracker{},
-		RPCService:             &RPCServiceMock{},
-		LedgerBackend:          mockLedgerBackend,
-		ChannelAccountStore:    mockChAccStore,
-		MetricsService:         mockMetricsService,
-		GetLedgersLimit:        defaultGetLedgersLimit,
-		Network:                network.TestNetworkPassphrase,
-		NetworkPassphrase:      network.TestNetworkPassphrase,
-		Archive:                &HistoryArchiveMock{},
-	})
-	require.NoError(t, err)
-
-	batch := BackfillBatch{StartLedger: 4599, EndLedger: 4599}
-	result := svc.processLedgersInBatchHistorical(ctx, mockLedgerBackend, batch)
-
-	require.NoError(t, result.Error)
-	assert.Equal(t, 1, result.LedgersCount)
-	assert.Nil(t, result.BatchChanges, "expected nil batch changes for historical mode")
-}
-
-// Test_ingestService_startCatchup_ProcessesBatchChanges tests the full catchup
-// flow including batch change processing and cursor updates.
-//
-//nolint:unparam // Test backend factory always returns nil error - this is intentional for testing
-func Test_ingestService_startCatchup_ProcessesBatchChanges(t *testing.T) {
-	dbt := dbtest.Open(t)
-	defer dbt.Close()
-	ctx := context.Background()
-	dbConnectionPool, err := db.OpenDBConnectionPool(ctx, dbt.DSN)
-	require.NoError(t, err)
-	defer dbConnectionPool.Close()
-
-	// Prepare a ledger metadata for the test
-	var ledgerMeta xdr.LedgerCloseMeta
-	err = xdr.SafeUnmarshalBase64(ledgerMetadataWith0Tx, &ledgerMeta)
-	require.NoError(t, err)
-	ledgerSeq := ledgerMeta.LedgerSequence()
-
-	// NOTE: processTokenChanges now only calls:
-	// 1. models.TrustlineAsset.BatchInsert (real DB operation)
-	// 2. contractMetadataService.FetchMetadata (mocked, but not triggered with empty data)
-	// 3. models.Contract.BatchInsert (real DB operation)
-	// 4. tokenIngestionService.ProcessTokenChanges(ctx, dbTx, trustlineChanges, contractChanges)
-	testCases := []struct {
-		name             string
-		setupMocks       func(t *testing.T, tokenIngestionService *TokenIngestionServiceMock, backendFactory *func(ctx context.Context) (ledgerbackend.LedgerBackend, error))
-		wantErr          bool
-		wantErrContains  string
-		wantLatestCursor uint32
-	}{
-		{
-			name: "successful_catchup_processes_token_changes",
-			setupMocks: func(t *testing.T, tokenIngestionService *TokenIngestionServiceMock, backendFactory *func(ctx context.Context) (ledgerbackend.LedgerBackend, error)) {
-				// ProcessTokenChanges is the only mock needed on tokenIngestionService
-				tokenIngestionService.On("ProcessTokenChanges", mock.Anything, mock.Anything, mock.Anything, mock.Anything, mock.Anything, mock.Anything).Return(nil)
-
-				// Backend factory
-				*backendFactory = func(_ context.Context) (ledgerbackend.LedgerBackend, error) {
-					mockBackend := &LedgerBackendMock{}
-					mockBackend.On("PrepareRange", mock.Anything, mock.Anything).Return(nil)
-					mockBackend.On("GetLedger", mock.Anything, ledgerSeq).Return(ledgerMeta, nil)
-					mockBackend.On("Close").Return(nil)
-					return mockBackend, nil
-				}
-			},
-			wantErr:          false,
-			wantLatestCursor: ledgerSeq,
-		},
-		{
-			name: "token_processing_error_returns_error",
-			setupMocks: func(t *testing.T, tokenIngestionService *TokenIngestionServiceMock, backendFactory *func(ctx context.Context) (ledgerbackend.LedgerBackend, error)) {
-				// ProcessTokenChanges fails
-				tokenIngestionService.On("ProcessTokenChanges", mock.Anything, mock.Anything, mock.Anything, mock.Anything, mock.Anything, mock.Anything).Return(fmt.Errorf("db connection error"))
-
-				// Backend factory
-				*backendFactory = func(_ context.Context) (ledgerbackend.LedgerBackend, error) {
-					mockBackend := &LedgerBackendMock{}
-					mockBackend.On("PrepareRange", mock.Anything, mock.Anything).Return(nil)
-					mockBackend.On("GetLedger", mock.Anything, ledgerSeq).Return(ledgerMeta, nil)
-					mockBackend.On("Close").Return(nil)
-					return mockBackend, nil
-				}
-			},
-			wantErr:          true,
-			wantErrContains:  "processing token changes",
-			wantLatestCursor: ledgerSeq - 1, // Cursor should NOT be updated
-		},
-		{
-			name: "cursor_not_updated_if_token_processing_fails",
-			setupMocks: func(t *testing.T, tokenIngestionService *TokenIngestionServiceMock, backendFactory *func(ctx context.Context) (ledgerbackend.LedgerBackend, error)) {
-				// ProcessTokenChanges fails - cursor should not be updated
-				tokenIngestionService.On("ProcessTokenChanges", mock.Anything, mock.Anything, mock.Anything, mock.Anything, mock.Anything, mock.Anything).Return(fmt.Errorf("db error"))
-
-				// Backend factory
-				*backendFactory = func(_ context.Context) (ledgerbackend.LedgerBackend, error) {
-					mockBackend := &LedgerBackendMock{}
-					mockBackend.On("PrepareRange", mock.Anything, mock.Anything).Return(nil)
-					mockBackend.On("GetLedger", mock.Anything, ledgerSeq).Return(ledgerMeta, nil)
-					mockBackend.On("Close").Return(nil)
-					return mockBackend, nil
-				}
-			},
-			wantErr:          true,
-			wantErrContains:  "processing token changes",
-			wantLatestCursor: ledgerSeq - 1, // Cursor should NOT be updated
-		},
-	}
-
-	for _, tc := range testCases {
-		t.Run(tc.name, func(t *testing.T) {
-			// Clean up and set up cursors (latest = startLedger - 1 for catchup mode validation)
-			setupDBCursors(t, ctx, dbConnectionPool, ledgerSeq-1, ledgerSeq-1)
-
-			mockMetricsService := metrics.NewMockMetricsService()
-			mockMetricsService.On("RegisterPoolMetrics", "ledger_indexer", mock.Anything).Return()
-			mockMetricsService.On("RegisterPoolMetrics", "backfill", mock.Anything).Return()
-			mockMetricsService.On("ObserveDBQueryDuration", mock.Anything, mock.Anything, mock.Anything).Return().Maybe()
-			mockMetricsService.On("IncDBQuery", mock.Anything, mock.Anything).Return().Maybe()
-			mockMetricsService.On("IncDBTransaction", mock.Anything).Return().Maybe()
-			mockMetricsService.On("ObserveDBTransactionDuration", mock.Anything, mock.Anything).Return().Maybe()
-			mockMetricsService.On("ObserveDBBatchSize", mock.Anything, mock.Anything, mock.Anything).Return().Maybe()
-			mockMetricsService.On("IncStateChanges", mock.Anything, mock.Anything, mock.Anything).Return().Maybe()
-			mockMetricsService.On("ObserveIngestionParticipantsCount", mock.Anything).Return().Maybe()
-			defer mockMetricsService.AssertExpectations(t)
-
-			models, err := data.NewModels(dbConnectionPool, mockMetricsService)
-			require.NoError(t, err)
-
-			mockRPCService := &RPCServiceMock{}
-			mockRPCService.On("NetworkPassphrase").Return(network.TestNetworkPassphrase).Maybe()
-
-			mockTokenIngestionService := NewTokenIngestionServiceMock(t)
-			var backendFactory func(ctx context.Context) (ledgerbackend.LedgerBackend, error)
-
-			tc.setupMocks(t, mockTokenIngestionService, &backendFactory)
-
-			svc, err := NewIngestService(IngestServiceConfig{
-				IngestionMode:          IngestionModeBackfill,
-				Models:                 models,
-				LatestLedgerCursorName: "latest_ledger_cursor",
-				OldestLedgerCursorName: "oldest_ledger_cursor",
-				AppTracker:             &apptracker.MockAppTracker{},
-				RPCService:             mockRPCService,
-				LedgerBackend:          &LedgerBackendMock{},
-				LedgerBackendFactory:   backendFactory,
-				ChannelAccountStore:    &store.ChannelAccountStoreMock{},
-				TokenIngestionService:  mockTokenIngestionService,
-				MetricsService:         mockMetricsService,
-				GetLedgersLimit:        defaultGetLedgersLimit,
-				Network:                network.TestNetworkPassphrase,
-				NetworkPassphrase:      network.TestNetworkPassphrase,
-				Archive:                &HistoryArchiveMock{},
-				BackfillBatchSize:      100,
-			})
-			require.NoError(t, err)
-
-			err = svc.startCatchup(ctx, ledgerSeq, ledgerSeq)
-
-			if tc.wantErr {
-				require.Error(t, err)
-				if tc.wantErrContains != "" {
-					assert.Contains(t, err.Error(), tc.wantErrContains)
-				}
-			} else {
-				require.NoError(t, err)
-			}
-
-			// Verify cursor state
-			cursor, err := models.IngestStore.Get(ctx, "latest_ledger_cursor")
-			require.NoError(t, err)
-			assert.Equal(t, tc.wantLatestCursor, cursor, "latest ledger cursor mismatch")
-		})
-	}
-}
-
-// Test_ingestService_processBackfillBatchesParallel_Historical verifies
-// that historical mode processes batches successfully.
-func Test_ingestService_processBackfillBatchesParallel_Historical(t *testing.T) {
-	dbt := dbtest.Open(t)
-	defer dbt.Close()
-	ctx := context.Background()
-	dbConnectionPool, err := db.OpenDBConnectionPool(ctx, dbt.DSN)
-	require.NoError(t, err)
-	defer dbConnectionPool.Close()
-
-	mockMetricsService := metrics.NewMockMetricsService()
-	mockMetricsService.On("RegisterPoolMetrics", "ledger_indexer", mock.Anything).Return()
-	mockMetricsService.On("RegisterPoolMetrics", "backfill", mock.Anything).Return()
-	mockMetricsService.On("SetOldestLedgerIngested", mock.Anything).Return().Maybe()
-	mockMetricsService.On("ObserveDBQueryDuration", mock.Anything, mock.Anything, mock.Anything).Return().Maybe()
-	mockMetricsService.On("IncDBQuery", mock.Anything, mock.Anything).Return().Maybe()
-	mockMetricsService.On("IncDBTransaction", mock.Anything).Return().Maybe()
-	mockMetricsService.On("ObserveDBTransactionDuration", mock.Anything, mock.Anything).Return().Maybe()
-	mockMetricsService.On("ObserveDBBatchSize", mock.Anything, mock.Anything, mock.Anything).Return().Maybe()
-	mockMetricsService.On("ObserveIngestionParticipantsCount", mock.Anything).Return().Maybe()
-	mockMetricsService.On("IncStateChanges", mock.Anything, mock.Anything, mock.Anything).Return().Maybe()
-	defer mockMetricsService.AssertExpectations(t)
-
-	models, modelsErr := data.NewModels(dbConnectionPool, mockMetricsService)
-	require.NoError(t, modelsErr)
-
-	factory := func(ctx context.Context) (ledgerbackend.LedgerBackend, error) {
-		mockBackend := &LedgerBackendMock{}
-		mockBackend.On("PrepareRange", mock.Anything, mock.Anything).Return(nil)
-		mockBackend.On("GetLedger", mock.Anything, mock.Anything).Return(xdr.LedgerCloseMeta{
-			V: 0,
-			V0: &xdr.LedgerCloseMetaV0{
-				LedgerHeader: xdr.LedgerHeaderHistoryEntry{
-					Header: xdr.LedgerHeader{
-						LedgerSeq: xdr.Uint32(100),
-					},
-				},
-			},
-		}, nil)
-		mockBackend.On("Close").Return(nil)
-		return mockBackend, nil
-	}
-
-	svc, svcErr := NewIngestService(IngestServiceConfig{
-		IngestionMode:          IngestionModeBackfill,
-		Models:                 models,
-		LatestLedgerCursorName: "latest_ledger_cursor",
-		OldestLedgerCursorName: "oldest_ledger_cursor",
-		AppTracker:             &apptracker.MockAppTracker{},
-		RPCService:             &RPCServiceMock{},
-		LedgerBackend:          &LedgerBackendMock{},
-		LedgerBackendFactory:   factory,
-		MetricsService:         mockMetricsService,
-		GetLedgersLimit:        defaultGetLedgersLimit,
-		Network:                network.TestNetworkPassphrase,
-		NetworkPassphrase:      network.TestNetworkPassphrase,
-		Archive:                &HistoryArchiveMock{},
-		BackfillBatchSize:      10,
-	})
-	require.NoError(t, svcErr)
-
-	batches := []BackfillBatch{
-		{StartLedger: 100, EndLedger: 100},
-		{StartLedger: 101, EndLedger: 101},
-	}
-
-	results := svc.processBackfillBatchesParallel(ctx, batches, svc.processLedgersInBatchHistorical, nil)
-
-	require.Len(t, results, 2)
-	for i, result := range results {
-		assert.NoError(t, result.Error, "batch %d should succeed", i)
-	}
-}
-
-// Test_ingestService_processBackfillBatchesParallel_Catchup verifies
-// that catchup mode processes batches successfully.
-func Test_ingestService_processBackfillBatchesParallel_Catchup(t *testing.T) {
-	dbt := dbtest.Open(t)
-	defer dbt.Close()
-	ctx := context.Background()
-	dbConnectionPool, err := db.OpenDBConnectionPool(ctx, dbt.DSN)
-	require.NoError(t, err)
-	defer dbConnectionPool.Close()
-
-	mockMetricsService := metrics.NewMockMetricsService()
-	mockMetricsService.On("RegisterPoolMetrics", "ledger_indexer", mock.Anything).Return()
-	mockMetricsService.On("RegisterPoolMetrics", "backfill", mock.Anything).Return()
-	mockMetricsService.On("ObserveDBQueryDuration", mock.Anything, mock.Anything, mock.Anything).Return().Maybe()
-	mockMetricsService.On("IncDBQuery", mock.Anything, mock.Anything).Return().Maybe()
-	mockMetricsService.On("IncDBTransaction", mock.Anything).Return().Maybe()
-	mockMetricsService.On("ObserveDBTransactionDuration", mock.Anything, mock.Anything).Return().Maybe()
-	mockMetricsService.On("ObserveDBBatchSize", mock.Anything, mock.Anything, mock.Anything).Return().Maybe()
-	mockMetricsService.On("ObserveIngestionParticipantsCount", mock.Anything).Return().Maybe()
-	mockMetricsService.On("IncStateChanges", mock.Anything, mock.Anything, mock.Anything).Return().Maybe()
-	defer mockMetricsService.AssertExpectations(t)
-
-	models, modelsErr := data.NewModels(dbConnectionPool, mockMetricsService)
-	require.NoError(t, modelsErr)
-
-	factory := func(ctx context.Context) (ledgerbackend.LedgerBackend, error) {
-		mockBackend := &LedgerBackendMock{}
-		mockBackend.On("PrepareRange", mock.Anything, mock.Anything).Return(nil)
-		mockBackend.On("GetLedger", mock.Anything, mock.Anything).Return(xdr.LedgerCloseMeta{
-			V: 0,
-			V0: &xdr.LedgerCloseMetaV0{
-				LedgerHeader: xdr.LedgerHeaderHistoryEntry{
-					Header: xdr.LedgerHeader{
-						LedgerSeq: xdr.Uint32(100),
-					},
-				},
-			},
-		}, nil)
-		mockBackend.On("Close").Return(nil)
-		return mockBackend, nil
-	}
-
-	svc, svcErr := NewIngestService(IngestServiceConfig{
-		IngestionMode:          IngestionModeBackfill,
-		Models:                 models,
-		LatestLedgerCursorName: "latest_ledger_cursor",
-		OldestLedgerCursorName: "oldest_ledger_cursor",
-		AppTracker:             &apptracker.MockAppTracker{},
-		RPCService:             &RPCServiceMock{},
-		LedgerBackend:          &LedgerBackendMock{},
-		LedgerBackendFactory:   factory,
-		MetricsService:         mockMetricsService,
-		GetLedgersLimit:        defaultGetLedgersLimit,
-		Network:                network.TestNetworkPassphrase,
-		NetworkPassphrase:      network.TestNetworkPassphrase,
-		Archive:                &HistoryArchiveMock{},
-		BackfillBatchSize:      10,
-	})
-	require.NoError(t, svcErr)
-
-	batches := []BackfillBatch{
-		{StartLedger: 100, EndLedger: 100},
-		{StartLedger: 101, EndLedger: 101},
-	}
-
-	results := svc.processBackfillBatchesParallel(ctx, batches, svc.processLedgersInBatchCatchup, nil)
-
-	require.Len(t, results, 2)
-	for i, result := range results {
-		assert.NoError(t, result.Error, "batch %d should succeed", i)
-	}
 }
